refactor(tools): split GUI verification modules
