+++ tools/dev/fluent_qt_gui_verify.py
@@ -10,387 +10,89 @@ new baseline or its own visual judgment.
 
 from __future__ import annotations
 
-import argparse
 from collections import Counter
-import copy
-from datetime import datetime, timezone
-from html import escape
-import hashlib
-import json
-import locale
-import math
-import os
+from datetime import datetime
 from pathlib import Path
+from typing import Any, Iterable, Mapping, Sequence
+import argparse
+import copy
+import locale
+import os
 import platform
 import re
 import shutil
-import struct
 import subprocess
 import sys
 import tempfile
-from typing import Any, Iterable, Mapping, Sequence
-import zlib
+
+# Resolve sibling modules for direct execution and file-based test loaders.
+sys.path.insert(0, str(Path(__file__).resolve().parent))
+
+from gui_verification.artifacts import (
+    canonical_json,
+    path_is_within,
+    paths_overlap,
+    png_dimensions,
+    read_json,
+    resolved_path,
+    sha256_bytes,
+    sha256_file,
+    write_json,
+)
+from gui_verification.common import (
+    ACTION_NAMES,
+    BASELINE_PROVENANCE_SCHEMA_VERSION,
+    BASELINE_SCHEMA_VERSION,
+    NATIVE_PLUGINS,
+    REVIEW_SCHEMA_VERSION,
+    SAFE_ID,
+    STATUS_PRIORITY,
+    TOOL_SCHEMA_VERSION,
+    VerificationError,
+    canonical_machine_name,
+    canonical_system_name,
+    check,
+    combined_status,
+    effective_require_native_desktop,
+    is_json_integer,
+    is_sha256,
+    is_trimmed_nonempty,
+    is_utc_timestamp,
+    merged_dict,
+    nested,
+    qt_scale_positive,
+    utc_now,
+    validated_device_pixel_ratio,
+)
+from gui_verification.comparison import (
+    comparator_policy_arguments,
+    geometry_baseline_check,
+    geometry_contract_check,
+    pixel_comparisons,
+    validate_comparator_report,
+)
+from gui_verification.presentation import (
+    review_html,
+)
+from gui_verification.recipe import (
+    validate_action_script,
+    validate_fields,
+    validate_geometry_policy,
+    validate_recipe,
+    validate_scenario_semantics,
+)
+from gui_verification.reports import (
+    baseline_report_errors,
+    capture_environment_check,
+    capture_environment_errors,
+    capture_report_check,
+    capture_report_errors,
+    inspector_report_errors,
+)
 
 
 PROJECT_ROOT = Path(__file__).resolve().parents[2]
 BUILD_TOOL = Path(__file__).with_name("fluent_qt_build.py")
-TOOL_SCHEMA_VERSION = 1
-RECIPE_SCHEMA_VERSION = 1
-BASELINE_SCHEMA_VERSION = 1
-BASELINE_PROVENANCE_SCHEMA_VERSION = 1
-REVIEW_SCHEMA_VERSION = 1
-SAFE_ID = re.compile(r"^[a-z0-9][a-z0-9._-]*$")
-TAG = re.compile(r"^[a-z][a-z0-9_-]*:[a-z0-9][a-z0-9._-]*$")
-SUPPORTED_TAG_PREFIXES = {
-    "theme",
-    "width",
-    "state",
-    "input",
-    "direction",
-    "platform",
-}
-NATIVE_PLUGINS = {"cocoa", "windows", "xcb", "wayland", "wayland-egl"}
-MAX_GEOMETRY_TOLERANCE = 32
-MAX_INSPECTOR_BUDGET = 100
-MAX_CHANNEL_THRESHOLD = 64
-MAX_DIFFERENT_RATIO = 0.10
-MAX_DEVICE_PIXEL_RATIO = 16.0
-MAX_PIXEL_COORDINATE = 32768
-MAX_SETTLE_MS = 10_000
-MAX_TIMEOUT_SECONDS = 300
-MAX_JSON_INTEGER_DIGITS = 4096
-MAX_JSON_NESTING_DEPTH = 128
-ACTION_NAMES = {
-    "focus",
-    "click",
-    "mouse_move",
-    "mouse_press",
-    "mouse_release",
-    "mouse_leave",
-    "key",
-    "type_text",
-    "set_property",
-    "wait",
-}
-KEYBOARD_ACTIONS = {"key", "type_text"}
-POINTER_ACTIONS = {
-    "click",
-    "mouse_move",
-    "mouse_press",
-    "mouse_release",
-    "mouse_leave",
-}
-STATUS_PRIORITY = {
-    "pass": 0,
-    "not-applicable": 0,
-    "human-required": 1,
-    "review-required": 1,
-    "incomplete": 2,
-    "fail": 3,
-}
-
-
-class VerificationError(RuntimeError):
-    """Raised for invalid inputs or unusable verification infrastructure."""
-
-
-def utc_now() -> str:
-    return datetime.now(timezone.utc).replace(microsecond=0).isoformat()
-
-
-def sha256_bytes(payload: bytes) -> str:
-    return hashlib.sha256(payload).hexdigest()
-
-
-def sha256_file(path: Path) -> str:
-    digest = hashlib.sha256()
-    with path.open("rb") as stream:
-        for chunk in iter(lambda: stream.read(1024 * 1024), b""):
-            digest.update(chunk)
-    return digest.hexdigest()
-
-
-def is_sha256(value: object) -> bool:
-    return isinstance(value, str) and re.fullmatch(r"[0-9a-f]{64}", value) is not None
-
-
-def is_trimmed_nonempty(value: object) -> bool:
-    return isinstance(value, str) and bool(value) and value == value.strip()
-
-
-def is_utc_timestamp(value: object) -> bool:
-    if not is_trimmed_nonempty(value):
-        return False
-    try:
-        parsed = datetime.fromisoformat(value)
-    except ValueError:
-        return False
-    return parsed.tzinfo is not None and parsed.utcoffset() == timezone.utc.utcoffset(
-        parsed
-    )
-
-
-def qt_round_positive(value: float) -> int | None:
-    """Match qRound() for the non-negative geometry values used here."""
-
-    try:
-        numeric = float(value)
-    except (OverflowError, TypeError, ValueError):
-        return None
-    if not math.isfinite(numeric) or numeric < 0:
-        return None
-    return math.floor(numeric + 0.5)
-
-
-def qt_scale_positive(value: object, scale: object) -> int | None:
-    try:
-        scaled = float(value) * float(scale)
-    except (OverflowError, TypeError, ValueError):
-        return None
-    return qt_round_positive(scaled)
-
-
-def png_dimensions(path: Path) -> tuple[int, int] | None:
-    try:
-        payload = path.read_bytes()
-    except OSError:
-        return None
-    if len(payload) < 8 or payload[:8] != b"\x89PNG\r\n\x1a\n":
-        return None
-
-    offset = 8
-    ihdr: bytes | None = None
-    idat_chunks: list[bytes] = []
-    seen_plte = False
-    seen_idat = False
-    idat_closed = False
-    seen_iend = False
-    while offset < len(payload):
-        if offset + 12 > len(payload):
-            return None
-        length = struct.unpack(">I", payload[offset : offset + 4])[0]
-        chunk_end = offset + 12 + length
-        if chunk_end > len(payload):
-            return None
-        chunk_type = payload[offset + 4 : offset + 8]
-        if (
-            re.fullmatch(rb"[A-Za-z]{4}", chunk_type) is None
-            or not 65 <= chunk_type[2] <= 90
-        ):
-            return None
-        chunk_payload = payload[offset + 8 : offset + 8 + length]
-        recorded_crc = struct.unpack(">I", payload[offset + 8 + length : chunk_end])[0]
-        computed_crc = zlib.crc32(chunk_type)
-        computed_crc = zlib.crc32(chunk_payload, computed_crc) & 0xFFFFFFFF
-        if recorded_crc != computed_crc:
-            return None
-        if ihdr is None and chunk_type != b"IHDR":
-            return None
-        if chunk_type == b"IHDR":
-            if ihdr is not None or length != 13 or offset != 8:
-                return None
-            ihdr = chunk_payload
-        elif chunk_type == b"PLTE":
-            if seen_idat or seen_plte or not length or length % 3 or length > 768:
-                return None
-            seen_plte = True
-        elif chunk_type == b"IDAT":
-            if ihdr is None or idat_closed or seen_iend:
-                return None
-            seen_idat = True
-            idat_chunks.append(chunk_payload)
-        elif chunk_type == b"IEND":
-            if length or not seen_idat or seen_iend:
-                return None
-            seen_iend = True
-            offset = chunk_end
-            break
-        else:
-            if seen_idat:
-                idat_closed = True
-            if chunk_type and 65 <= chunk_type[0] <= 90:
-                return None
-        offset = chunk_end
-
-    if ihdr is None or not seen_iend or offset != len(payload):
-        return None
-    width, height, bit_depth, color_type, compression, filtering, interlace = (
-        struct.unpack(">IIBBBBB", ihdr)
-    )
-    valid_depths = {
-        0: {1, 2, 4, 8, 16},
-        2: {8, 16},
-        3: {1, 2, 4, 8},
-        4: {8, 16},
-        6: {8, 16},
-    }
-    if (
-        width <= 0
-        or height <= 0
-        or width > 32768
-        or height > 32768
-        or color_type not in valid_depths
-        or bit_depth not in valid_depths[color_type]
-        or compression != 0
-        or filtering != 0
-        or interlace not in {0, 1}
-        or (color_type == 3 and not seen_plte)
-        or (color_type in {0, 4} and seen_plte)
-    ):
-        return None
-
-    samples = {0: 1, 2: 3, 3: 1, 4: 2, 6: 4}[color_type]
-
-    def row_size(row_width: int) -> int:
-        return 1 + ((row_width * samples * bit_depth + 7) // 8)
-
-    scanline_sizes: list[int] = []
-    if interlace == 0:
-        scanline_sizes = [row_size(width)] * height
-    else:
-        for start_x, start_y, step_x, step_y in (
-            (0, 0, 8, 8),
-            (4, 0, 8, 8),
-            (0, 4, 4, 8),
-            (2, 0, 4, 4),
-            (0, 2, 2, 4),
-            (1, 0, 2, 2),
-            (0, 1, 1, 2),
-        ):
-            pass_width = max(0, (width - start_x + step_x - 1) // step_x)
-            pass_height = max(0, (height - start_y + step_y - 1) // step_y)
-            if pass_width:
-                scanline_sizes.extend([row_size(pass_width)] * pass_height)
-    expected_bytes = sum(scanline_sizes)
-    if expected_bytes <= 0 or expected_bytes > 256 * 1024 * 1024:
-        return None
-
-    filter_offsets: list[int] = []
-    scanline_offset = 0
-    for size in scanline_sizes:
-        filter_offsets.append(scanline_offset)
-        scanline_offset += size
-    filter_index = 0
-    decoded_bytes = 0
-
-    def consume(decoded: bytes) -> bool:
-        nonlocal decoded_bytes, filter_index
-        if decoded_bytes + len(decoded) > expected_bytes:
-            return False
-        end = decoded_bytes + len(decoded)
-        while filter_index < len(filter_offsets) and filter_offsets[filter_index] < end:
-            position = filter_offsets[filter_index] - decoded_bytes
-            if position >= 0 and decoded[position] > 4:
-                return False
-            filter_index += 1
-        decoded_bytes = end
-        return True
-
-    try:
-        decoder = zlib.decompressobj()
-        for compressed in idat_chunks:
-            remaining = compressed
-            while remaining:
-                decoded = decoder.decompress(remaining, 1024 * 1024)
-                remaining = decoder.unconsumed_tail
-                if not consume(decoded):
-                    return None
-                if not decoded and not remaining:
-                    break
-        if not consume(decoder.flush()):
-            return None
-    except zlib.error:
-        return None
-    if (
-        not decoder.eof
-        or decoder.unused_data
-        or decoded_bytes != expected_bytes
-        or filter_index != len(filter_offsets)
-    ):
-        return None
-    return width, height
-
-
-def canonical_json(value: object) -> bytes:
-    return json.dumps(
-        value, ensure_ascii=False, sort_keys=True, separators=(",", ":")
-    ).encode("utf-8")
-
-
-def parse_json_integer(raw: str) -> int:
-    digits = raw[1:] if raw.startswith("-") else raw
-    if len(digits) > MAX_JSON_INTEGER_DIGITS:
-        raise ValueError(
-            f"JSON integer exceeds {MAX_JSON_INTEGER_DIGITS} digits"
-        )
-    return int(raw)
-
-
-def validate_json_nesting(value: object) -> None:
-    pending: list[tuple[object, int]] = [(value, 1)]
-    while pending:
-        current, depth = pending.pop()
-        if depth > MAX_JSON_NESTING_DEPTH:
-            raise ValueError(
-                f"JSON nesting exceeds {MAX_JSON_NESTING_DEPTH} levels"
-            )
-        if isinstance(current, dict):
-            pending.extend((item, depth + 1) for item in current.values())
-        elif isinstance(current, list):
-            pending.extend((item, depth + 1) for item in current)
-
-
-def read_json(path: Path) -> dict[str, Any]:
-    try:
-        value = json.loads(
-            path.read_text(encoding="utf-8"), parse_int=parse_json_integer
-        )
-        validate_json_nesting(value)
-    except OSError as error:
-        raise VerificationError(f"Could not read JSON {path}: {error}") from error
-    except (ValueError, RecursionError) as error:
-        raise VerificationError(f"Could not parse JSON {path}: {error}") from error
-    if not isinstance(value, dict):
-        raise VerificationError(f"JSON root must be an object: {path}")
-    return value
-
-
-def write_json(path: Path, value: object) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    temporary: Path | None = None
-    try:
-        with tempfile.NamedTemporaryFile(
-            mode="w",
-            encoding="utf-8",
-            dir=path.parent,
-            prefix=f".{path.name}.",
-            suffix=".tmp",
-            delete=False,
-        ) as stream:
-            json.dump(value, stream, ensure_ascii=False, indent=2)
-            stream.write("\n")
-            temporary = Path(stream.name)
-        temporary.replace(path)
-    finally:
-        if temporary is not None and temporary.exists():
-            temporary.unlink()
-
-
-def resolved_path(raw: str | os.PathLike[str], base: Path) -> Path:
-    path = Path(raw).expanduser()
-    return (base / path).resolve() if not path.is_absolute() else path.resolve()
-
-
-def path_is_within(path: Path, directory: Path) -> bool:
-    try:
-        path.resolve().relative_to(directory.resolve())
-    except ValueError:
-        return False
-    return True
-
-
-def paths_overlap(first: Path, second: Path) -> bool:
-    return path_is_within(first, second) or path_is_within(second, first)
 
 
 def approved_baseline_root() -> Path:
@@ -419,126 +121,6 @@ def recipe_path_base(recipe: Mapping[str, object], recipe_path: Path) -> Path:
     raise VerificationError("path_base must be repository or recipe")
 
 
-def merged_dict(*values: object) -> dict[str, Any]:
-    result: dict[str, Any] = {}
-    for value in values:
-        if isinstance(value, dict):
-            result.update(value)
-    return result
-
-
-def effective_require_native_desktop(
-    defaults: Mapping[str, object], scenario: Mapping[str, object]
-) -> bool:
-    if "require_native_desktop" in scenario:
-        value = scenario["require_native_desktop"]
-    elif "require_native_desktop" in defaults:
-        value = defaults["require_native_desktop"]
-    else:
-        value = True
-    return value if isinstance(value, bool) else True
-
-
-def nested(value: object, *keys: str) -> object:
-    current = value
-    for key in keys:
-        if not isinstance(current, dict):
-            return None
-        current = current.get(key)
-    return current
-
-
-def validate_fields(
-    value: Mapping[str, object],
-    allowed: set[str],
-    context: str,
-    errors: list[str],
-    required: set[str] | None = None,
-) -> None:
-    unknown = sorted(set(value) - allowed)
-    missing = sorted((required or set()) - set(value))
-    if unknown:
-        errors.append(f"{context} has unsupported fields: {', '.join(unknown)}")
-    if missing:
-        errors.append(f"{context} is missing fields: {', '.join(missing)}")
-
-
-def validate_inspector_policy(
-    value: object, context: str, errors: list[str]
-) -> None:
-    if not isinstance(value, dict):
-        errors.append(f"{context} must be an object")
-        return
-    validate_fields(
-        value,
-        {"max_findings", "max_by_severity", "allowed_codes"},
-        context,
-        errors,
-    )
-    maximum = value.get("max_findings")
-    if "max_findings" in value and (
-        not isinstance(maximum, int)
-        or isinstance(maximum, bool)
-        or not 0 <= maximum <= MAX_INSPECTOR_BUDGET
-    ):
-        errors.append(
-            f"{context}.max_findings must be from 0 to {MAX_INSPECTOR_BUDGET}"
-        )
-    severity = value.get("max_by_severity")
-    if "max_by_severity" in value:
-        if not isinstance(severity, dict):
-            errors.append(f"{context}.max_by_severity must be an object")
-        else:
-            validate_fields(
-                severity, {"info", "warning", "error"},
-                f"{context}.max_by_severity", errors
-            )
-            for name, budget in severity.items():
-                if (
-                    not isinstance(budget, int)
-                    or isinstance(budget, bool)
-                    or not 0 <= budget <= MAX_INSPECTOR_BUDGET
-                ):
-                    errors.append(
-                        f"{context}.max_by_severity.{name} must be from 0 to "
-                        f"{MAX_INSPECTOR_BUDGET}"
-                    )
-    allowed_codes = value.get("allowed_codes")
-    if "allowed_codes" in value and (
-        not isinstance(allowed_codes, list)
-        or not all(isinstance(code, str) and code for code in allowed_codes)
-        or len(set(allowed_codes)) != len(allowed_codes)
-    ):
-        errors.append(f"{context}.allowed_codes must be a unique string array")
-
-
-def check(
-    check_id: str,
-    status: str,
-    message: str,
-    details: object | None = None,
-) -> dict[str, object]:
-    result: dict[str, object] = {
-        "id": check_id,
-        "status": status,
-        "message": message,
-    }
-    if details is not None:
-        result["details"] = details
-    return result
-
-
-def combined_status(checks: Sequence[Mapping[str, object]]) -> str:
-    status = "pass"
-    for item in checks:
-        candidate = str(item.get("status", "incomplete"))
-        if STATUS_PRIORITY.get(candidate, STATUS_PRIORITY["incomplete"]) > STATUS_PRIORITY[
-            status
-        ]:
-            status = candidate
-    return status
-
-
 def default_preset() -> str:
     system = platform.system().lower()
     machine = platform.machine().lower()
@@ -549,871 +131,6 @@ def default_preset() -> str:
     if system == "linux":
         return "vcpkg-linux-arm64" if machine in {"arm64", "aarch64"} else "vcpkg-linux"
     raise VerificationError(f"Unsupported GUI verification host: {platform.system()}")
-
-
-def validate_size(value: object, context: str, errors: list[str]) -> None:
-    if not isinstance(value, str) or not re.fullmatch(r"[1-9][0-9]*x[1-9][0-9]*", value):
-        errors.append(f"{context} must be WIDTHxHEIGHT")
-        return
-    width, height = (int(piece) for piece in value.split("x", 1))
-    if not 320 <= width <= 3840 or not 240 <= height <= 2160:
-        errors.append(f"{context} must be within 320x240 and 3840x2160")
-
-
-def validate_runtime_options(
-    value: Mapping[str, object], context: str, errors: list[str]
-) -> None:
-    settle_ms = value.get("settle_ms")
-    if "settle_ms" in value and (
-        not isinstance(settle_ms, int)
-        or isinstance(settle_ms, bool)
-        or not 0 <= settle_ms <= MAX_SETTLE_MS
-    ):
-        errors.append(f"{context}.settle_ms must be from 0 to {MAX_SETTLE_MS}")
-    timeout = value.get("timeout_seconds")
-    if "timeout_seconds" in value and (
-        not isinstance(timeout, int)
-        or isinstance(timeout, bool)
-        or not 1 <= timeout <= MAX_TIMEOUT_SECONDS
-    ):
-        errors.append(
-            f"{context}.timeout_seconds must be from 1 to {MAX_TIMEOUT_SECONDS}"
-        )
-    native = value.get("require_native_desktop")
-    if "require_native_desktop" in value and not isinstance(native, bool):
-        errors.append(f"{context}.require_native_desktop must be a boolean")
-
-
-def validate_environment(value: object, context: str, errors: list[str]) -> None:
-    if not isinstance(value, dict):
-        errors.append(f"{context} must be an object")
-        return
-    for name, setting in value.items():
-        if not isinstance(name, str) or not name:
-            errors.append(f"{context} keys must be non-empty strings")
-        if not isinstance(setting, (str, int, float, bool)):
-            errors.append(
-                f"{context}.{name} must be a string, number, or boolean"
-            )
-
-
-def validate_geometry_policy(
-    value: object, context: str, errors: list[str], require_probes: bool
-) -> None:
-    if not isinstance(value, dict):
-        errors.append(f"{context} must be an object")
-        return
-    validate_fields(value, {"required", "tolerance"}, context, errors)
-    required = value.get("required")
-    if (require_probes or "required" in value) and (
-        not isinstance(required, list) or not required
-    ):
-        errors.append(f"{context}.required must be a non-empty array")
-    if isinstance(required, list):
-        probe_names: list[str] = []
-        for index, raw in enumerate(required):
-            if is_trimmed_nonempty(raw):
-                probe_names.append(raw)
-                continue
-            if not isinstance(raw, dict) or not is_trimmed_nonempty(
-                raw.get("object_name")
-            ):
-                errors.append(
-                    f"{context}.required[{index}] must name one object_name"
-                )
-                continue
-            probe_names.append(str(raw["object_name"]))
-            validate_fields(
-                raw,
-                {
-                    "object_name",
-                    "tolerance",
-                    "not_clipped",
-                    "min_width",
-                    "min_height",
-                    "max_width",
-                    "max_height",
-                    "rect",
-                },
-                f"{context}.required[{index}]",
-                errors,
-                {"object_name"},
-            )
-            probe_tolerance = raw.get("tolerance")
-            if "tolerance" in raw and (
-                not isinstance(probe_tolerance, int)
-                or isinstance(probe_tolerance, bool)
-                or not 0 <= probe_tolerance <= MAX_GEOMETRY_TOLERANCE
-            ):
-                errors.append(
-                    f"{context}.required[{index}].tolerance must be from 0 to "
-                    f"{MAX_GEOMETRY_TOLERANCE}"
-                )
-            rect = raw.get("rect")
-            if "rect" in raw and (
-                not isinstance(rect, dict)
-                or set(rect) != {"x", "y", "width", "height"}
-                or not all(
-                    isinstance(rect.get(key), int)
-                    and not isinstance(rect.get(key), bool)
-                    for key in rect
-                )
-                or int(rect.get("width", 0)) <= 0
-                or int(rect.get("height", 0)) <= 0
-            ):
-                errors.append(
-                    f"{context}.required[{index}].rect must contain integer x, y, width, and height"
-                )
-            not_clipped = raw.get("not_clipped")
-            if "not_clipped" in raw and not isinstance(not_clipped, bool):
-                errors.append(
-                    f"{context}.required[{index}].not_clipped must be a boolean"
-                )
-            for dimension in (
-                "min_width",
-                "min_height",
-                "max_width",
-                "max_height",
-            ):
-                candidate = raw.get(dimension)
-                if dimension in raw and (
-                    not isinstance(candidate, int)
-                    or isinstance(candidate, bool)
-                    or candidate < 1
-                ):
-                    errors.append(
-                        f"{context}.required[{index}].{dimension} must be a "
-                        "positive integer"
-                    )
-            for minimum, maximum in (
-                ("min_width", "max_width"),
-                ("min_height", "max_height"),
-            ):
-                lower = raw.get(minimum)
-                upper = raw.get(maximum)
-                if (
-                    isinstance(lower, int)
-                    and not isinstance(lower, bool)
-                    and isinstance(upper, int)
-                    and not isinstance(upper, bool)
-                    and lower > upper
-                ):
-                    errors.append(
-                        f"{context}.required[{index}].{minimum} must not exceed "
-                        f"{maximum}"
-                    )
-        if len(set(probe_names)) != len(probe_names):
-            errors.append(
-                f"{context}.required contains duplicate object_name probes"
-            )
-    tolerance = value.get("tolerance")
-    if "tolerance" in value and (
-        not isinstance(tolerance, int)
-        or isinstance(tolerance, bool)
-        or not 0 <= tolerance <= MAX_GEOMETRY_TOLERANCE
-    ):
-        errors.append(
-            f"{context}.tolerance must be from 0 to {MAX_GEOMETRY_TOLERANCE}"
-        )
-
-
-def validate_pixel_policy(
-    value: object,
-    context: str,
-    errors: list[str],
-    *,
-    allow_regions: bool = True,
-) -> None:
-    if not isinstance(value, dict):
-        errors.append(f"{context} must be an object")
-        return
-    allowed = {
-        "channel_threshold",
-        "max_different_pixels",
-        "max_different_ratio",
-        "search_radius",
-        "max_translation",
-        "edge_threshold",
-    }
-    if allow_regions:
-        allowed.add("regions")
-    validate_fields(value, allowed, context, errors)
-    integer_ranges = {
-        "channel_threshold": (0, 255),
-        "max_different_pixels": (0, None),
-        "search_radius": (0, 32),
-        "max_translation": (0, 32),
-        "edge_threshold": (1, 255),
-    }
-    for name, (minimum, maximum) in integer_ranges.items():
-        if name not in value:
-            continue
-        candidate = value[name]
-        if (
-            not isinstance(candidate, int)
-            or isinstance(candidate, bool)
-            or candidate < minimum
-            or (maximum is not None and candidate > maximum)
-        ):
-            errors.append(f"{context}.{name} is outside its valid range")
-    channel_threshold = value.get("channel_threshold")
-    if isinstance(channel_threshold, int) and channel_threshold > MAX_CHANNEL_THRESHOLD:
-        errors.append(
-            f"{context}.channel_threshold must not exceed {MAX_CHANNEL_THRESHOLD}"
-        )
-    ratio = value.get("max_different_ratio")
-    if "max_different_ratio" in value and (
-        not isinstance(ratio, (int, float))
-        or isinstance(ratio, bool)
-        or not 0 <= ratio <= MAX_DIFFERENT_RATIO
-    ):
-        errors.append(
-            f"{context}.max_different_ratio must be from 0 to "
-            f"{MAX_DIFFERENT_RATIO}"
-        )
-    if "regions" not in value:
-        return
-    regions = value.get("regions")
-    if not allow_regions:
-        errors.append(f"{context}.regions is not allowed in a region override")
-        return
-    if not isinstance(regions, list):
-        errors.append(f"{context}.regions must be an array")
-        return
-    region_ids: set[str] = set()
-    for index, raw in enumerate(regions):
-        if not isinstance(raw, dict):
-            errors.append(f"{context}.regions[{index}] must be an object")
-            continue
-        validate_fields(
-            raw,
-            {"id", "rect", "coordinate_space", "policy"},
-            f"{context}.regions[{index}]",
-            errors,
-            {"id", "rect"},
-        )
-        region_id = raw.get("id")
-        if not isinstance(region_id, str) or not SAFE_ID.fullmatch(region_id):
-            errors.append(f"{context}.regions[{index}].id is invalid")
-        elif region_id in region_ids:
-            errors.append(f"{context}.regions[{index}].id is duplicated")
-        else:
-            region_ids.add(region_id)
-        rect = raw.get("rect")
-        if (
-            not isinstance(rect, list)
-            or len(rect) != 4
-            or not all(
-                isinstance(item, int) and not isinstance(item, bool)
-                for item in rect
-            )
-            or (isinstance(rect, list) and len(rect) == 4 and (rect[0] < 0 or rect[1] < 0 or rect[2] <= 0 or rect[3] <= 0))
-        ):
-            errors.append(
-                f"{context}.regions[{index}].rect must be non-negative x,y and positive width,height"
-            )
-        elif any(item > MAX_PIXEL_COORDINATE for item in rect):
-            errors.append(
-                f"{context}.regions[{index}].rect values must not exceed "
-                f"{MAX_PIXEL_COORDINATE}"
-            )
-        if raw.get("coordinate_space", "logical") not in {"logical", "device"}:
-            errors.append(
-                f"{context}.regions[{index}].coordinate_space must be logical or device"
-            )
-        if "policy" in raw:
-            validate_pixel_policy(
-                raw["policy"],
-                f"{context}.regions[{index}].policy",
-                errors,
-                allow_regions=False,
-            )
-            region_policy = raw.get("policy")
-            max_pixels = (
-                region_policy.get("max_different_pixels")
-                if isinstance(region_policy, dict)
-                else None
-            )
-            if (
-                isinstance(rect, list)
-                and len(rect) == 4
-                and all(
-                    isinstance(item, int) and not isinstance(item, bool)
-                    for item in rect
-                )
-                and isinstance(max_pixels, int)
-                and not isinstance(max_pixels, bool)
-                and max_pixels > (rect[2] * rect[3]) // 10
-            ):
-                errors.append(
-                    f"{context}.regions[{index}].policy.max_different_pixels "
-                    f"must not exceed {MAX_DIFFERENT_RATIO:.0%} of the region pixels"
-                )
-
-
-def validate_action_script(
-    value: object, context: str, errors: list[str]
-) -> dict[str, object] | None:
-    if not isinstance(value, dict):
-        errors.append(f"{context} must be a JSON object")
-        return None
-    validate_fields(
-        value,
-        {"schema_version", "stop_on_failure", "steps"},
-        context,
-        errors,
-        {"schema_version", "steps"},
-    )
-    if value.get("schema_version") != 1:
-        errors.append(f"{context}.schema_version must be 1")
-    if "stop_on_failure" in value and not isinstance(
-        value.get("stop_on_failure"), bool
-    ):
-        errors.append(f"{context}.stop_on_failure must be a boolean")
-    steps = value.get("steps")
-    if not isinstance(steps, list) or not steps:
-        errors.append(f"{context}.steps must be a non-empty array")
-        return value
-    ids: set[str] = set()
-    allowed_step_fields = {
-        "id",
-        "action",
-        "target",
-        "descendant_class",
-        "position",
-        "button",
-        "modifiers",
-        "key",
-        "text",
-        "property",
-        "value",
-        "milliseconds",
-        "after_ms",
-        "observe",
-        "expect",
-    }
-    for index, step in enumerate(steps):
-        step_context = f"{context}.steps[{index}]"
-        if not isinstance(step, dict):
-            errors.append(f"{step_context} must be an object")
-            continue
-        validate_fields(
-            step, allowed_step_fields, step_context, errors, {"action"}
-        )
-        action = step.get("action")
-        if action not in ACTION_NAMES:
-            errors.append(f"{step_context}.action is unsupported")
-        step_id = step.get("id")
-        if "id" in step:
-            if not isinstance(step_id, str) or not step_id:
-                errors.append(f"{step_context}.id must be a non-empty string")
-            elif step_id in ids:
-                errors.append(f"{step_context}.id is duplicated")
-            else:
-                ids.add(step_id)
-        for name in ("target", "descendant_class", "key", "property"):
-            candidate = step.get(name)
-            if name in step and (
-                not isinstance(candidate, str) or not candidate
-            ):
-                errors.append(f"{step_context}.{name} must be a non-empty string")
-        if "text" in step and not isinstance(step.get("text"), str):
-            errors.append(f"{step_context}.text must be a string")
-        if action == "type_text" and (
-            not isinstance(step.get("text"), str) or not step.get("text")
-        ):
-            errors.append(f"{step_context}.type_text requires non-empty text")
-        position = step.get("position")
-        if "position" in step and (
-            not isinstance(position, dict)
-            or set(position) != {"x", "y"}
-            or not all(
-                isinstance(position.get(name), int)
-                and not isinstance(position.get(name), bool)
-                for name in ("x", "y")
-            )
-        ):
-            errors.append(f"{step_context}.position must contain integer x and y")
-        if "button" in step and step.get("button") not in {
-            "left",
-            "right",
-            "middle",
-        }:
-            errors.append(f"{step_context}.button is unsupported")
-        modifiers = step.get("modifiers")
-        if "modifiers" in step and (
-            not isinstance(modifiers, list)
-            or len(set(modifiers)) != len(modifiers)
-            or not all(
-                modifier in {"shift", "control", "ctrl", "alt", "meta", "shortcut"}
-                for modifier in modifiers
-            )
-        ):
-            errors.append(f"{step_context}.modifiers is invalid")
-        for name in ("milliseconds", "after_ms"):
-            candidate = step.get(name)
-            if name in step and (
-                not isinstance(candidate, int)
-                or isinstance(candidate, bool)
-                or not 0 <= candidate <= MAX_SETTLE_MS
-            ):
-                errors.append(
-                    f"{step_context}.{name} must be from 0 to {MAX_SETTLE_MS}"
-                )
-        observe = step.get("observe")
-        if "observe" in step and (
-            not isinstance(observe, list)
-            or len(set(observe)) != len(observe)
-            or not all(isinstance(item, str) and item for item in observe)
-        ):
-            errors.append(f"{step_context}.observe must be a unique string array")
-        expectations = step.get("expect")
-        if "expect" in step and (
-            not isinstance(expectations, dict) or not expectations
-        ):
-            errors.append(f"{step_context}.expect must be a non-empty object")
-    return value
-
-
-def validate_scenario_semantics(
-    scenario: Mapping[str, object],
-    action_script: Mapping[str, object] | None,
-    context: str,
-    errors: list[str],
-    *,
-    external_actions_pending: bool = False,
-    require_native_desktop: bool = True,
-) -> None:
-    tags = scenario.get("tags")
-    if not isinstance(tags, list) or not all(isinstance(tag, str) for tag in tags):
-        return
-    if len(set(tags)) != len(tags):
-        errors.append(f"{context}.tags must not contain duplicates")
-    malformed = [tag for tag in tags if not TAG.fullmatch(tag)]
-    if malformed:
-        errors.append(f"{context}.tags contain malformed values: {', '.join(malformed)}")
-    unsupported_prefixes = sorted(
-        {
-            tag.split(":", 1)[0]
-            for tag in tags
-            if ":" in tag and tag.split(":", 1)[0] not in SUPPORTED_TAG_PREFIXES
-        }
-    )
-    if unsupported_prefixes:
-        errors.append(
-            f"{context}.tags use unsupported coverage categories: "
-            + ", ".join(unsupported_prefixes)
-        )
-    theme_tags = [tag for tag in tags if tag.startswith("theme:")]
-    expected_theme = f"theme:{scenario.get('theme')}"
-    if theme_tags != [expected_theme]:
-        errors.append(f"{context}.tags must contain exactly {expected_theme}")
-    direction_tags = [tag for tag in tags if tag.startswith("direction:")]
-    if direction_tags and direction_tags != [
-        f"direction:{scenario.get('direction', 'ltr')}"
-    ]:
-        errors.append(f"{context}.direction tag does not match the scenario")
-    width_tags = [tag for tag in tags if tag.startswith("width:")]
-    if len(width_tags) != 1:
-        errors.append(f"{context}.tags must contain exactly one width tag")
-    else:
-        size = scenario.get("size")
-        if isinstance(size, str) and "x" in size:
-            try:
-                width = int(size.split("x", 1)[0])
-            except ValueError:
-                width = 0
-            expected_width = (
-                "width:narrow"
-                if width <= 640
-                else "width:normal"
-                if width <= 1007
-                else "width:wide"
-            )
-            if width_tags[0] != expected_width:
-                errors.append(
-                    f"{context}.{width_tags[0]} does not match viewport width {width}"
-                )
-    platform_tags = [tag for tag in tags if tag.startswith("platform:")]
-    unsupported_platform_tags = sorted(
-        set(platform_tags) - {"platform:desktop-qpa"}
-    )
-    if unsupported_platform_tags:
-        errors.append(
-            f"{context} has unsupported platform tags: "
-            + ", ".join(unsupported_platform_tags)
-        )
-    if "platform:desktop-qpa" in tags and not require_native_desktop:
-        errors.append(
-            f"{context} platform:desktop-qpa cannot disable desktop QPA capture"
-        )
-    native_state_tags = sorted(tag for tag in tags if tag.startswith("state:native-"))
-    if native_state_tags:
-        errors.append(
-            f"{context} cannot claim native state coverage from the Gallery runner: "
-            + ", ".join(native_state_tags)
-        )
-    if external_actions_pending:
-        return
-    input_tags = [tag for tag in tags if tag.startswith("input:")]
-    unknown_input_tags = sorted(set(input_tags) - {"input:keyboard", "input:mouse"})
-    if unknown_input_tags:
-        errors.append(
-            f"{context} has unsupported input tags: {', '.join(unknown_input_tags)}"
-        )
-    state_tags = [tag for tag in tags if tag.startswith("state:")]
-    supported_state_tags = {
-        "state:default",
-        "state:focus",
-        "state:hover",
-        "state:pressed",
-    }
-    unknown_state_tags = sorted(set(state_tags) - supported_state_tags)
-    if unknown_state_tags:
-        errors.append(
-            f"{context} has unsupported state tags: "
-            + ", ".join(unknown_state_tags)
-        )
-    if len(state_tags) != 1:
-        errors.append(f"{context} must declare exactly one state tag")
-    steps = (
-        action_script.get("steps")
-        if isinstance(action_script, Mapping)
-        and isinstance(action_script.get("steps"), list)
-        else []
-    )
-    actions = {
-        step.get("action")
-        for step in steps
-        if isinstance(step, dict) and isinstance(step.get("action"), str)
-    }
-    has_assertion = any(
-        isinstance(step, dict)
-        and isinstance(step.get("expect"), dict)
-        and bool(step.get("expect"))
-        for step in steps
-    )
-    if input_tags and action_script is None:
-        errors.append(f"{context} input coverage requires an action script")
-    if "input:keyboard" in input_tags and not (actions & KEYBOARD_ACTIONS):
-        errors.append(f"{context} keyboard coverage requires a key or type_text action")
-    if "input:mouse" in input_tags and not (actions & POINTER_ACTIONS):
-        errors.append(f"{context} mouse coverage requires a pointer action")
-    if input_tags and not has_assertion:
-        errors.append(f"{context} input coverage requires a non-empty expectation")
-    non_default_states = [tag for tag in state_tags if tag != "state:default"]
-    if non_default_states and action_script is None:
-        errors.append(f"{context} non-default state coverage requires an action script")
-    if non_default_states and not has_assertion:
-        errors.append(
-            f"{context} non-default state coverage requires a non-empty expectation"
-        )
-    if "state:focus" in state_tags and not (
-        actions & (KEYBOARD_ACTIONS | {"focus"})
-    ):
-        errors.append(f"{context} focus state requires focus or keyboard input")
-    final_step = steps[-1] if steps and isinstance(steps[-1], dict) else {}
-    final_action = final_step.get("action")
-    final_expectation = (
-        final_step.get("expect")
-        if isinstance(final_step.get("expect"), dict)
-        else {}
-    )
-    if "state:focus" in state_tags and final_expectation.get("has_focus") is not True:
-        errors.append(
-            f"{context} focus state requires a final has_focus=true assertion"
-        )
-    if "state:hover" in state_tags and "mouse_move" not in actions:
-        errors.append(f"{context} hover state requires mouse_move input")
-    if "state:hover" in state_tags and (
-        final_action != "mouse_move" or not final_expectation
-    ):
-        errors.append(
-            f"{context} hover state must end with an asserted mouse_move action"
-        )
-    if "state:pressed" in state_tags and "mouse_press" not in actions:
-        errors.append(f"{context} pressed state requires mouse_press input")
-    if "state:pressed" in state_tags and (
-        final_action != "mouse_press" or not final_expectation
-    ):
-        errors.append(
-            f"{context} pressed state must end with an asserted mouse_press action"
-        )
-
-
-def validate_recipe(recipe: Mapping[str, object]) -> list[str]:
-    errors: list[str] = []
-    validate_fields(
-        recipe,
-        {
-            "schema_version",
-            "id",
-            "path_base",
-            "author",
-            "selection",
-            "coverage",
-            "environment",
-            "defaults",
-            "scenarios",
-        },
-        "recipe",
-        errors,
-        {
-            "schema_version",
-            "id",
-            "path_base",
-            "author",
-            "selection",
-            "coverage",
-            "defaults",
-            "scenarios",
-        },
-    )
-    if recipe.get("schema_version") != RECIPE_SCHEMA_VERSION:
-        errors.append("schema_version must be 1")
-    recipe_id = recipe.get("id")
-    if not isinstance(recipe_id, str) or not SAFE_ID.fullmatch(recipe_id):
-        errors.append("id must use lowercase letters, numbers, dot, underscore, or dash")
-    if recipe.get("path_base") not in {"repository", "recipe"}:
-        errors.append("path_base must be repository or recipe")
-    author = recipe.get("author")
-    if not isinstance(author, dict) or not is_trimmed_nonempty(author.get("id")):
-        errors.append("author.id is required")
-    elif author.get("kind") not in {"ai", "human"}:
-        errors.append("author.kind must be ai or human")
-    if isinstance(author, dict):
-        validate_fields(
-            author, {"id", "kind"}, "author", errors, {"id", "kind"}
-        )
-    selection = recipe.get("selection")
-    if not isinstance(selection, dict) or not isinstance(selection.get("route"), str) or not selection.get("route"):
-        errors.append("selection.route is required")
-    if isinstance(selection, dict):
-        validate_fields(
-            selection, {"route", "sample"}, "selection", errors, {"route"}
-        )
-        if "sample" in selection and not is_trimmed_nonempty(
-            selection.get("sample")
-        ):
-            errors.append("selection.sample must be a non-empty string")
-
-    coverage = recipe.get("coverage")
-    required_tags = coverage.get("required_tags") if isinstance(coverage, dict) else None
-    if not isinstance(required_tags, list) or not required_tags or not all(
-        isinstance(tag, str) and tag for tag in required_tags
-    ):
-        errors.append("coverage.required_tags must be a non-empty string array")
-    elif len(set(required_tags)) != len(required_tags):
-        errors.append("coverage.required_tags must not contain duplicates")
-    elif any(not TAG.fullmatch(tag) for tag in required_tags):
-        errors.append("coverage.required_tags contain malformed values")
-    if isinstance(coverage, dict):
-        validate_fields(
-            coverage,
-            {"required_tags"},
-            "coverage",
-            errors,
-            {"required_tags"},
-        )
-
-    if "environment" in recipe:
-        validate_environment(recipe.get("environment"), "environment", errors)
-
-    defaults = recipe.get("defaults")
-    defaults = defaults if isinstance(defaults, dict) else {}
-    validate_fields(
-        defaults,
-        {
-            "settle_ms",
-            "timeout_seconds",
-            "require_native_desktop",
-            "inspector",
-            "geometry",
-            "pixel",
-        },
-        "defaults",
-        errors,
-        {"inspector", "geometry", "pixel"},
-    )
-    validate_runtime_options(defaults, "defaults", errors)
-    validate_inspector_policy(defaults.get("inspector"), "defaults.inspector", errors)
-    validate_geometry_policy(
-        defaults.get("geometry"), "defaults.geometry", errors, True
-    )
-    validate_pixel_policy(defaults.get("pixel"), "defaults.pixel", errors)
-
-    scenarios = recipe.get("scenarios")
-    if not isinstance(scenarios, list) or not scenarios:
-        errors.append("scenarios must be a non-empty array")
-        return errors
-    seen: set[str] = set()
-    covered: set[str] = set()
-    for index, raw in enumerate(scenarios):
-        context = f"scenarios[{index}]"
-        if not isinstance(raw, dict):
-            errors.append(f"{context} must be an object")
-            continue
-        validate_fields(
-            raw,
-            {
-                "id",
-                "selection",
-                "theme",
-                "direction",
-                "size",
-                "tags",
-                "baseline",
-                "review",
-                "actions",
-                "environment",
-                "settle_ms",
-                "timeout_seconds",
-                "require_native_desktop",
-                "inspector",
-                "geometry",
-                "pixel",
-            },
-            context,
-            errors,
-            {"id", "theme", "size", "tags", "baseline", "review"},
-        )
-        scenario_id = raw.get("id")
-        if not isinstance(scenario_id, str) or not SAFE_ID.fullmatch(scenario_id):
-            errors.append(f"{context}.id is invalid")
-        elif scenario_id in seen:
-            errors.append(f"{context}.id is duplicated: {scenario_id}")
-        else:
-            seen.add(scenario_id)
-        if raw.get("theme") not in {"light", "dark"}:
-            errors.append(f"{context}.theme must be light or dark")
-        if raw.get("direction", "ltr") not in {"ltr", "rtl"}:
-            errors.append(f"{context}.direction must be ltr or rtl")
-        validate_size(raw.get("size"), f"{context}.size", errors)
-        tags = raw.get("tags")
-        if not isinstance(tags, list) or not tags or not all(
-            isinstance(tag, str) and tag for tag in tags
-        ):
-            errors.append(f"{context}.tags must be a non-empty string array")
-        else:
-            covered.update(tags)
-        scenario_selection = raw.get("selection")
-        if "selection" in raw:
-            if not isinstance(scenario_selection, dict):
-                errors.append(f"{context}.selection must be an object")
-            else:
-                validate_fields(
-                    scenario_selection,
-                    {"route", "sample"},
-                    f"{context}.selection",
-                    errors,
-                    {"route"},
-                )
-                if not is_trimmed_nonempty(scenario_selection.get("route")):
-                    errors.append(
-                        f"{context}.selection.route must be a non-empty string"
-                    )
-                if "sample" in scenario_selection and not is_trimmed_nonempty(
-                    scenario_selection.get("sample")
-                ):
-                    errors.append(
-                        f"{context}.selection.sample must be a non-empty string"
-                    )
-        baseline = raw.get("baseline")
-        if not isinstance(baseline, (str, dict)):
-            errors.append(f"{context}.baseline must be a path or platform map")
-        elif isinstance(baseline, str) and not baseline.strip():
-            errors.append(f"{context}.baseline must not be empty")
-        elif isinstance(baseline, dict) and (
-            not baseline
-            or not all(
-                isinstance(key, str)
-                and key
-                and isinstance(path, str)
-                and path
-                for key, path in baseline.items()
-            )
-        ):
-            errors.append(
-                f"{context}.baseline platform map must contain non-empty paths"
-            )
-        review = raw.get("review")
-        if not isinstance(review, list) or not review or not all(
-            is_trimmed_nonempty(item) for item in review
-        ):
-            errors.append(f"{context}.review must contain visual review prompts")
-        actions = raw.get("actions")
-        if "actions" in raw and not isinstance(actions, (str, dict)):
-            errors.append(f"{context}.actions must be a path or JSON object")
-        inline_actions = None
-        if isinstance(actions, dict):
-            inline_actions = validate_action_script(
-                actions, f"{context}.actions", errors
-            )
-        validate_scenario_semantics(
-            raw,
-            inline_actions,
-            context,
-            errors,
-            external_actions_pending=isinstance(actions, str),
-            require_native_desktop=effective_require_native_desktop(
-                defaults, raw
-            ),
-        )
-        if "environment" in raw:
-            validate_environment(
-                raw.get("environment"), f"{context}.environment", errors
-            )
-        validate_runtime_options(raw, context, errors)
-        if "inspector" in raw:
-            validate_inspector_policy(
-                raw.get("inspector"), f"{context}.inspector", errors
-            )
-        if "geometry" in raw:
-            validate_geometry_policy(raw.get("geometry"), f"{context}.geometry", errors, False)
-        if "pixel" in raw:
-            validate_pixel_policy(raw.get("pixel"), f"{context}.pixel", errors)
-        size = raw.get("size")
-        if isinstance(size, str) and re.fullmatch(r"[1-9][0-9]*x[1-9][0-9]*", size):
-            width, height = (int(piece) for piece in size.split("x", 1))
-            pixel_policy = merged_dict(defaults.get("pixel"), raw.get("pixel"))
-            regions = pixel_policy.get("regions")
-            if isinstance(regions, list):
-                for region_index, region in enumerate(regions):
-                    if not isinstance(region, dict):
-                        continue
-                    rect = region.get("rect")
-                    if (
-                        region.get("coordinate_space", "logical") == "logical"
-                        and isinstance(rect, list)
-                        and len(rect) == 4
-                        and all(
-                            isinstance(item, int) and not isinstance(item, bool)
-                            for item in rect
-                        )
-                        and (
-                            rect[0] + rect[2] > width
-                            or rect[1] + rect[3] > height
-                        )
-                    ):
-                        errors.append(
-                            f"{context}.pixel.regions[{region_index}] exceeds "
-                            "the logical viewport"
-                        )
-            max_pixels = pixel_policy.get("max_different_pixels")
-            if (
-                isinstance(max_pixels, int)
-                and not isinstance(max_pixels, bool)
-                and max_pixels > (width * height) // 10
-            ):
-                errors.append(
-                    f"{context}.pixel.max_different_pixels must not exceed "
-                    f"{MAX_DIFFERENT_RATIO:.0%} of the logical viewport pixels"
-                )
-    if isinstance(required_tags, list):
-        missing = sorted(set(required_tags) - covered)
-        if missing:
-            errors.append("coverage is missing required tags: " + ", ".join(missing))
-    return errors
 
 
 def configured_build_dir(args: argparse.Namespace) -> Path:
@@ -1547,29 +264,6 @@ def git_state() -> dict[str, object]:
         "revision": revision.stdout.strip() if revision.returncode == 0 else None,
         "dirty": bool(status.stdout.strip()) if status.returncode == 0 else None,
     }
-
-
-def canonical_system_name(value: object) -> str:
-    raw = value.lower() if isinstance(value, str) else ""
-    return {
-        "darwin": "macos",
-        "macos": "macos",
-        "osx": "macos",
-        "windows": "windows",
-        "winnt": "windows",
-        "linux": "linux",
-    }.get(raw, raw)
-
-
-def canonical_machine_name(value: object) -> str:
-    raw = value.lower() if isinstance(value, str) else ""
-    return {
-        "aarch64": "arm64",
-        "arm64": "arm64",
-        "amd64": "x64",
-        "x86_64": "x64",
-        "x64": "x64",
-    }.get(raw, raw)
 
 
 def host_key() -> tuple[str, str]:
@@ -1950,1052 +644,6 @@ def inspector_check(report: Mapping[str, object], policy: Mapping[str, object]) 
             "remaining_findings": relevant,
             "counts": dict(by_severity),
         },
-    )
-
-
-def is_json_integer(
-    value: object,
-    *,
-    minimum: int | None = None,
-    maximum: int | None = None,
-) -> bool:
-    return (
-        isinstance(value, int)
-        and not isinstance(value, bool)
-        and (minimum is None or value >= minimum)
-        and (maximum is None or value <= maximum)
-    )
-
-
-def is_json_number(
-    value: object,
-    *,
-    minimum: float | None = None,
-    maximum: float | None = None,
-) -> bool:
-    if not isinstance(value, (int, float)) or isinstance(value, bool):
-        return False
-    try:
-        numeric = float(value)
-    except (OverflowError, TypeError, ValueError):
-        return False
-    return (
-        math.isfinite(numeric)
-        and (minimum is None or numeric >= minimum)
-        and (maximum is None or numeric <= maximum)
-    )
-
-
-def validated_device_pixel_ratio(value: object) -> float | None:
-    if not is_json_number(
-        value,
-        minimum=0.000001,
-        maximum=MAX_DEVICE_PIXEL_RATIO,
-    ):
-        return None
-    try:
-        return float(value)
-    except (OverflowError, TypeError, ValueError):
-        return None
-
-
-def object_fields_are(
-    value: object,
-    required: set[str],
-    allowed: set[str] | None = None,
-) -> bool:
-    return (
-        isinstance(value, dict)
-        and required <= set(value)
-        and (allowed is None or set(value) <= allowed)
-    )
-
-
-def rect_report_errors(value: object, context: str) -> list[str]:
-    fields = {"x", "y", "width", "height"}
-    if not object_fields_are(value, fields, fields):
-        return [f"{context} must contain exactly x, y, width, and height"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    for name in ("x", "y"):
-        if not is_json_integer(value.get(name)):
-            errors.append(f"{context}.{name} must be an integer")
-    for name in ("width", "height"):
-        if not is_json_integer(value.get(name), minimum=0):
-            errors.append(f"{context}.{name} must be a non-negative integer")
-    return errors
-
-
-def size_report_errors(
-    value: object, context: str, *, minimum: int | None = 0
-) -> list[str]:
-    fields = {"width", "height"}
-    if not object_fields_are(value, fields, fields):
-        return [f"{context} must contain exactly width and height"]
-    assert isinstance(value, dict)
-    return [
-        f"{context}.{name} must be an integer"
-        + ("" if minimum is None else f" >= {minimum}")
-        for name in ("width", "height")
-        if not is_json_integer(value.get(name), minimum=minimum)
-    ]
-
-
-def capture_environment_errors(value: object) -> list[str]:
-    fields = {
-        "fingerprint_schema_version",
-        "qt_version",
-        "platform_plugin",
-        "style",
-        "device_pixel_ratio",
-        "logical_dpi_x",
-        "logical_dpi_y",
-        "locale",
-        "font",
-        "screen",
-        "system",
-        "scale_environment",
-    }
-    if not object_fields_are(value, fields, fields):
-        return ["capture environment has missing or unsupported top-level fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("fingerprint_schema_version") != 1:
-        errors.append("capture environment fingerprint_schema_version must be 1")
-    for name in ("qt_version", "platform_plugin", "style", "locale"):
-        if not isinstance(value.get(name), str) or not value.get(name):
-            errors.append(f"capture environment {name} must be non-empty")
-    if validated_device_pixel_ratio(value.get("device_pixel_ratio")) is None:
-        errors.append(
-            "capture environment device_pixel_ratio must be positive and <= "
-            f"{MAX_DEVICE_PIXEL_RATIO:g}"
-        )
-    for name in ("logical_dpi_x", "logical_dpi_y"):
-        if not is_json_number(value.get(name), minimum=0.000001):
-            errors.append(f"capture environment {name} must be positive")
-
-    font_fields = {
-        "family",
-        "style_name",
-        "point_size",
-        "pixel_size",
-        "weight",
-        "italic",
-    }
-    font = value.get("font")
-    if not object_fields_are(font, font_fields, font_fields):
-        errors.append("capture environment font has missing or unsupported fields")
-    else:
-        assert isinstance(font, dict)
-        if not isinstance(font.get("family"), str) or not font.get("family"):
-            errors.append("capture environment font.family must be non-empty")
-        if not isinstance(font.get("style_name"), str):
-            errors.append("capture environment font.style_name must be a string")
-        if not is_json_number(font.get("point_size")):
-            errors.append("capture environment font.point_size must be numeric")
-        for name in ("pixel_size", "weight"):
-            if not is_json_integer(font.get(name)):
-                errors.append(f"capture environment font.{name} must be an integer")
-        if not isinstance(font.get("italic"), bool):
-            errors.append("capture environment font.italic must be a boolean")
-
-    screen_fields = {
-        "depth",
-        "geometry",
-        "available_geometry",
-        "physical_dpi_x",
-        "physical_dpi_y",
-    }
-    screen_fields |= {"name", "manufacturer", "model", "serial_number"}
-    screen = value.get("screen")
-    if not object_fields_are(screen, screen_fields, screen_fields):
-        errors.append("capture environment screen has missing or unsupported fields")
-    else:
-        assert isinstance(screen, dict)
-        for name in ("name", "manufacturer", "model", "serial_number"):
-            if not isinstance(screen.get(name), str):
-                errors.append(f"capture environment screen.{name} must be a string")
-        if not is_json_integer(screen.get("depth"), minimum=1):
-            errors.append("capture environment screen.depth must be positive")
-        for name in ("geometry", "available_geometry"):
-            errors.extend(
-                rect_report_errors(
-                    screen.get(name), f"capture environment screen.{name}"
-                )
-            )
-            rect = screen.get(name)
-            if isinstance(rect, dict) and any(
-                not is_json_integer(rect.get(dimension), minimum=1)
-                for dimension in ("width", "height")
-            ):
-                errors.append(
-                    f"capture environment screen.{name} must have positive area"
-                )
-        for name in ("physical_dpi_x", "physical_dpi_y"):
-            if not is_json_number(screen.get(name), minimum=0.000001):
-                errors.append(f"capture environment screen.{name} must be positive")
-
-    system_fields = {
-        "product_type",
-        "product_version",
-        "kernel_type",
-        "kernel_version",
-        "cpu_architecture",
-    }
-    system = value.get("system")
-    if not object_fields_are(system, system_fields, system_fields):
-        errors.append("capture environment system has missing or unsupported fields")
-    else:
-        assert isinstance(system, dict)
-        for name in system_fields:
-            if not isinstance(system.get(name), str):
-                errors.append(f"capture environment system.{name} must be a string")
-        for name in ("product_type", "kernel_type", "cpu_architecture"):
-            if isinstance(system.get(name), str) and not system.get(name):
-                errors.append(f"capture environment system.{name} must be non-empty")
-
-    scale_fields = {
-        "QT_SCALE_FACTOR",
-        "QT_SCREEN_SCALE_FACTORS",
-        "QT_FONT_DPI",
-        "QT_AUTO_SCREEN_SCALE_FACTOR",
-        "QT_ENABLE_HIGHDPI_SCALING",
-    }
-    scale_environment = value.get("scale_environment")
-    if not object_fields_are(scale_environment, scale_fields, scale_fields):
-        errors.append(
-            "capture environment scale_environment has missing or unsupported fields"
-        )
-    elif not all(
-        isinstance(scale_environment.get(name), str) for name in scale_fields
-    ):
-        errors.append("capture environment scale_environment values must be strings")
-    return errors
-
-
-def capture_environment_check(report: Mapping[str, object]) -> dict[str, object]:
-    errors = capture_environment_errors(report.get("environment"))
-    return check(
-        "capture.environment",
-        "incomplete" if errors else "pass",
-        "Capture environment fingerprint is complete and well formed."
-        if not errors
-        else "Capture environment fingerprint is incomplete or malformed.",
-        {"validation_errors": errors},
-    )
-
-
-def interaction_report_errors(value: object) -> list[str]:
-    required = {"schema_version", "requested", "status", "summary", "steps"}
-    allowed = required | {"source", "error"}
-    if not object_fields_are(value, required, allowed):
-        return ["interaction_report has missing or unsupported fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("interaction_report.schema_version must be 1")
-    if not isinstance(value.get("requested"), bool):
-        errors.append("interaction_report.requested must be a boolean")
-    if value.get("status") not in {"pass", "fail", "not-requested"}:
-        errors.append("interaction_report.status is unsupported")
-    if "source" in value and not isinstance(value.get("source"), str):
-        errors.append("interaction_report.source must be a string")
-    if "error" in value and not isinstance(value.get("error"), str):
-        errors.append("interaction_report.error must be a string")
-    summary_fields = {"total", "executed", "passed", "failed"}
-    summary = value.get("summary")
-    if not object_fields_are(summary, summary_fields, summary_fields):
-        errors.append("interaction_report.summary is malformed")
-    else:
-        assert isinstance(summary, dict)
-        if any(
-            not is_json_integer(summary.get(name), minimum=0)
-            for name in summary_fields
-        ):
-            errors.append("interaction_report.summary values must be integers")
-    steps = value.get("steps")
-    if not isinstance(steps, list):
-        return [*errors, "interaction_report.steps must be an array"]
-    requested = value.get("requested")
-    if requested is True:
-        if not is_trimmed_nonempty(value.get("source")):
-            errors.append(
-                "requested interaction_report.source must be non-empty"
-            )
-        if value.get("status") not in {"pass", "fail"}:
-            errors.append("requested interaction_report status is invalid")
-    elif requested is False:
-        if "source" in value:
-            errors.append(
-                "not-requested interaction_report must not contain source"
-            )
-        if value.get("status") != "not-requested" or steps:
-            errors.append("not-requested interaction_report is inconsistent")
-    if isinstance(summary, dict) and all(
-        is_json_integer(summary.get(name), minimum=0)
-        for name in summary_fields
-    ):
-        if summary.get("executed") != len(steps):
-            errors.append(
-                "interaction_report.summary.executed does not match steps"
-            )
-        if summary.get("passed") + summary.get("failed") != len(steps):
-            errors.append(
-                "interaction_report.summary outcomes do not match steps"
-            )
-        if summary.get("total") < summary.get("executed"):
-            errors.append(
-                "interaction_report.summary.total is smaller than executed"
-            )
-    step_fields = {
-        "index",
-        "request",
-        "id",
-        "action",
-        "target",
-        "mechanism",
-        "expect",
-        "observation",
-        "status",
-        "message",
-    }
-    for index, step in enumerate(steps):
-        context = f"interaction_report.steps[{index}]"
-        if not isinstance(step, dict) or not {"index", "status"} <= set(step):
-            errors.append(f"{context} is malformed")
-            continue
-        if not set(step) <= step_fields:
-            errors.append(f"{context} has unsupported fields")
-        if not is_json_integer(step.get("index"), minimum=0):
-            errors.append(f"{context}.index must be a non-negative integer")
-        if step.get("status") not in {"pass", "fail"}:
-            errors.append(f"{context}.status is unsupported")
-        for name in ("request", "expect", "observation"):
-            if name in step and not isinstance(step.get(name), dict):
-                errors.append(f"{context}.{name} must be an object")
-        for name in ("id", "action", "target", "mechanism", "message"):
-            if name in step and not isinstance(step.get(name), str):
-                errors.append(f"{context}.{name} must be a string")
-    return errors
-
-
-def capture_report_errors(value: object) -> list[str]:
-    fields = {
-        "schema_version",
-        "tool",
-        "status",
-        "selection",
-        "scene",
-        "environment",
-        "artifacts",
-        "interaction_report",
-        "geometry_report",
-        "quality_report",
-    }
-    if not object_fields_are(value, fields, fields):
-        return ["capture report has missing or unsupported top-level fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 2:
-        errors.append("capture report schema_version must be 2")
-    if value.get("tool") != "FluentQt Gallery Preview":
-        errors.append("capture report tool is unsupported")
-    if value.get("status") not in {"ok", "artifact-error", "interaction-error"}:
-        errors.append("capture report status is unsupported")
-
-    selection_fields = {"route", "sample"}
-    selection = value.get("selection")
-    if not object_fields_are(selection, selection_fields, selection_fields):
-        errors.append("capture report selection is malformed")
-    else:
-        assert isinstance(selection, dict)
-        if not isinstance(selection.get("route"), str) or not selection.get("route"):
-            errors.append("capture report selection.route must be non-empty")
-        if not isinstance(selection.get("sample"), str):
-            errors.append("capture report selection.sample must be a string")
-
-    scene_fields = {
-        "requested_theme",
-        "theme",
-        "layout_direction",
-        "settle_ms",
-        "requested_width",
-        "requested_height",
-        "actual_width",
-        "actual_height",
-    }
-    scene = value.get("scene")
-    if not object_fields_are(scene, scene_fields, scene_fields):
-        errors.append("capture report scene is malformed")
-    else:
-        assert isinstance(scene, dict)
-        for name in ("requested_theme", "theme"):
-            if scene.get(name) not in {"light", "dark"}:
-                errors.append(f"capture report scene.{name} is unsupported")
-        if scene.get("layout_direction") not in {"ltr", "rtl"}:
-            errors.append("capture report scene.layout_direction is unsupported")
-        if not is_json_integer(scene.get("settle_ms"), minimum=0):
-            errors.append("capture report scene.settle_ms must be non-negative")
-        for name in ("requested_width", "actual_width"):
-            if not is_json_integer(scene.get(name), minimum=1, maximum=3840):
-                errors.append(
-                    f"capture report scene.{name} must be from 1 to 3840"
-                )
-        for name in ("requested_height", "actual_height"):
-            if not is_json_integer(scene.get(name), minimum=1, maximum=2160):
-                errors.append(
-                    f"capture report scene.{name} must be from 1 to 2160"
-                )
-
-    artifacts = value.get("artifacts")
-    if not object_fields_are(artifacts, {"snapshot"}, {"snapshot"}):
-        errors.append("capture report artifacts is malformed")
-    else:
-        assert isinstance(artifacts, dict)
-        snapshot_fields = {"requested", "written", "path", "sha256", "error"}
-        snapshot = artifacts.get("snapshot")
-        if not object_fields_are(snapshot, snapshot_fields, snapshot_fields):
-            errors.append("capture report artifacts.snapshot is malformed")
-        else:
-            assert isinstance(snapshot, dict)
-            for name in ("requested", "written"):
-                if not isinstance(snapshot.get(name), bool):
-                    errors.append(
-                        f"capture report artifacts.snapshot.{name} must be a boolean"
-                    )
-            for name in ("path", "error"):
-                if not isinstance(snapshot.get(name), str):
-                    errors.append(
-                        f"capture report artifacts.snapshot.{name} must be a string"
-                    )
-            if not isinstance(snapshot.get("sha256"), str):
-                errors.append(
-                    "capture report artifacts.snapshot.sha256 must be a string"
-                )
-            if snapshot.get("written") is True and not is_sha256(
-                snapshot.get("sha256")
-            ):
-                errors.append(
-                    "capture report artifacts.snapshot.sha256 must be a digest"
-                )
-            if snapshot.get("written") is True and (
-                snapshot.get("requested") is not True
-                or not snapshot.get("path")
-                or bool(snapshot.get("error"))
-            ):
-                errors.append(
-                    "capture report written snapshot state is inconsistent"
-                )
-    errors.extend(capture_environment_errors(value.get("environment")))
-    errors.extend(interaction_report_errors(value.get("interaction_report")))
-    errors.extend(geometry_report_errors(value.get("geometry_report")))
-    errors.extend(inspector_report_errors(value.get("quality_report"), value))
-    return errors
-
-
-def capture_report_check(report: Mapping[str, object]) -> dict[str, object]:
-    errors = capture_report_errors(report)
-    return check(
-        "capture.contract",
-        "incomplete" if errors else "pass",
-        "Capture report is closed-schema and internally valid."
-        if not errors
-        else "Capture report is malformed or contains unsupported fields.",
-        {"validation_errors": errors},
-    )
-
-
-def inspector_report_errors(
-    value: object, capture_report: Mapping[str, object] | None = None
-) -> list[str]:
-    fields = {"schema_version", "tool", "root", "summary", "findings"}
-    if not object_fields_are(value, fields, fields):
-        return ["quality_report has missing or unsupported top-level fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("quality_report.schema_version must be 1")
-    if value.get("tool") != "FluentQt Inspector":
-        errors.append("quality_report.tool is unsupported")
-    root = value.get("root")
-    root_fields = {"class", "object_name", "width", "height"}
-    if not object_fields_are(root, root_fields, root_fields):
-        errors.append("quality_report.root is malformed")
-    else:
-        assert isinstance(root, dict)
-        if not isinstance(root.get("class"), str) or not root.get("class"):
-            errors.append("quality_report.root.class must be non-empty")
-        if not isinstance(root.get("object_name"), str):
-            errors.append("quality_report.root.object_name must be a string")
-        for name in ("width", "height"):
-            if not is_json_integer(root.get(name), minimum=0):
-                errors.append(
-                    f"quality_report.root.{name} must be a non-negative integer"
-                )
-        if capture_report is not None:
-            expected_sizes = [
-                (
-                    nested(capture_report, "scene", "actual_width"),
-                    nested(capture_report, "scene", "actual_height"),
-                    "scene",
-                ),
-                (
-                    nested(capture_report, "geometry_report", "root_size", "width"),
-                    nested(capture_report, "geometry_report", "root_size", "height"),
-                    "geometry_report",
-                ),
-            ]
-            for width, height, source in expected_sizes:
-                if (
-                    is_json_integer(width, minimum=0)
-                    and is_json_integer(height, minimum=0)
-                    and (root.get("width"), root.get("height")) != (width, height)
-                ):
-                    errors.append(
-                        f"quality_report.root size does not match {source}"
-                    )
-    findings = value.get("findings")
-    if not isinstance(findings, list):
-        return [*errors, "quality_report.findings must be an array"]
-    valid_severities = {"info", "warning", "error"}
-    valid_categories = {
-        "text",
-        "accessibility",
-        "input",
-        "focus",
-        "layout",
-        "actions",
-        "scrolling",
-    }
-    finding_fields = {
-        "code", "category", "severity", "path", "rect", "message", "details"
-    }
-    severity_counts: Counter[str] = Counter()
-    category_counts: Counter[str] = Counter()
-    for index, finding in enumerate(findings):
-        context = f"quality_report.findings[{index}]"
-        if not object_fields_are(finding, finding_fields, finding_fields):
-            errors.append(f"{context} has missing or unsupported fields")
-            continue
-        assert isinstance(finding, dict)
-        code = finding.get("code")
-        if not isinstance(code, str) or re.fullmatch(
-            r"[a-z]+(?:[.-][a-z]+)*", code
-        ) is None:
-            errors.append(f"{context}.code is malformed")
-        category = finding.get("category")
-        if category not in valid_categories:
-            errors.append(f"{context}.category is unsupported")
-        else:
-            category_counts[str(category)] += 1
-        severity = finding.get("severity")
-        if severity not in valid_severities:
-            errors.append(f"{context}.severity is unsupported")
-        else:
-            severity_counts[str(severity)] += 1
-        if not isinstance(finding.get("path"), str) or not finding.get("path"):
-            errors.append(f"{context}.path must be non-empty")
-        errors.extend(rect_report_errors(finding.get("rect"), f"{context}.rect"))
-        if not isinstance(finding.get("message"), str) or not finding.get("message"):
-            errors.append(f"{context}.message must be non-empty")
-        if not isinstance(finding.get("details"), dict):
-            errors.append(f"{context}.details must be an object")
-    summary = value.get("summary")
-    summary_fields = {"findings", "by_severity", "by_category"}
-    if not object_fields_are(summary, summary_fields, summary_fields):
-        errors.append("quality_report.summary is malformed")
-        return errors
-    assert isinstance(summary, dict)
-    if summary.get("findings") != len(findings):
-        errors.append("quality_report.summary.findings does not match findings")
-    by_severity = summary.get("by_severity")
-    if not object_fields_are(by_severity, valid_severities, valid_severities):
-        errors.append("quality_report.summary.by_severity is malformed")
-    else:
-        assert isinstance(by_severity, dict)
-        for severity in sorted(valid_severities):
-            if not is_json_integer(by_severity.get(severity), minimum=0):
-                errors.append(
-                    f"quality_report.summary.by_severity.{severity} is invalid"
-                )
-            elif by_severity.get(severity) != severity_counts[severity]:
-                errors.append(
-                    f"quality_report.summary.by_severity.{severity} does not match findings"
-                )
-    by_category = summary.get("by_category")
-    if not isinstance(by_category, dict) or not all(
-        category in valid_categories and is_json_integer(count, minimum=0)
-        for category, count in by_category.items()
-    ):
-        errors.append("quality_report.summary.by_category is malformed")
-    elif dict(category_counts) != by_category:
-        errors.append("quality_report.summary.by_category does not match findings")
-    return errors
-
-
-def geometry_report_errors(value: object) -> list[str]:
-    fields = {"schema_version", "tool", "root_size", "widget_count", "widgets"}
-    if not object_fields_are(value, fields, fields):
-        return ["geometry_report has missing or unsupported top-level fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("geometry_report.schema_version must be 1")
-    if value.get("tool") != "FluentQt Named Widget Geometry":
-        errors.append("geometry_report.tool is unsupported")
-    errors.extend(size_report_errors(value.get("root_size"), "geometry_report.root_size"))
-    widgets = value.get("widgets")
-    if not isinstance(widgets, list):
-        return [*errors, "geometry_report.widgets must be an array"]
-    if value.get("widget_count") != len(widgets):
-        errors.append("geometry_report.widget_count does not match widgets")
-    widget_fields = {
-        "path", "class", "object_name", "stable", "rect", "visible_rect",
-        "minimum_size", "maximum_size", "size_hint", "enabled", "has_focus",
-        "clipped", "layout_direction", "accessible_name",
-    }
-    for index, widget in enumerate(widgets):
-        context = f"geometry_report.widgets[{index}]"
-        if not object_fields_are(widget, widget_fields, widget_fields):
-            errors.append(f"{context} has missing or unsupported fields")
-            continue
-        assert isinstance(widget, dict)
-        for name in ("path", "class"):
-            if not isinstance(widget.get(name), str) or not widget.get(name):
-                errors.append(f"{context}.{name} must be non-empty")
-        for name in ("object_name", "accessible_name"):
-            if not isinstance(widget.get(name), str):
-                errors.append(f"{context}.{name} must be a string")
-        if widget.get("stable") is not bool(widget.get("object_name")):
-            errors.append(f"{context}.stable does not match object_name")
-        for name in ("enabled", "has_focus", "clipped"):
-            if not isinstance(widget.get(name), bool):
-                errors.append(f"{context}.{name} must be a boolean")
-        if widget.get("layout_direction") not in {"ltr", "rtl"}:
-            errors.append(f"{context}.layout_direction is unsupported")
-        for name in ("rect", "visible_rect"):
-            errors.extend(rect_report_errors(widget.get(name), f"{context}.{name}"))
-        rect = widget.get("rect")
-        visible_rect = widget.get("visible_rect")
-        if isinstance(rect, dict) and isinstance(visible_rect, dict):
-            derived_clipped = visible_rect != rect
-            if widget.get("clipped") is not derived_clipped:
-                errors.append(f"{context}.clipped does not match visible_rect")
-            if all(
-                is_json_integer(rect.get(name))
-                and is_json_integer(visible_rect.get(name))
-                for name in ("x", "y", "width", "height")
-            ) and visible_rect.get("width", 0) > 0 and visible_rect.get("height", 0) > 0:
-                rect_right = int(rect["x"]) + int(rect["width"])
-                rect_bottom = int(rect["y"]) + int(rect["height"])
-                visible_right = int(visible_rect["x"]) + int(visible_rect["width"])
-                visible_bottom = int(visible_rect["y"]) + int(visible_rect["height"])
-                root_size = value.get("root_size")
-                root_width = int(root_size.get("width", 0)) if isinstance(root_size, dict) else 0
-                root_height = int(root_size.get("height", 0)) if isinstance(root_size, dict) else 0
-                if not (
-                    int(rect["x"]) <= int(visible_rect["x"]) <= visible_right <= rect_right
-                    and int(rect["y"]) <= int(visible_rect["y"]) <= visible_bottom <= rect_bottom
-                    and 0 <= int(visible_rect["x"]) <= visible_right <= root_width
-                    and 0 <= int(visible_rect["y"]) <= visible_bottom <= root_height
-                ):
-                    errors.append(
-                        f"{context}.visible_rect is outside rect or root_size"
-                    )
-        for name in ("minimum_size", "maximum_size"):
-            errors.extend(size_report_errors(widget.get(name), f"{context}.{name}"))
-        errors.extend(
-            size_report_errors(
-                widget.get("size_hint"), f"{context}.size_hint", minimum=None
-            )
-        )
-    return errors
-
-
-def baseline_geometry_report_errors(value: object) -> list[str]:
-    fields = {"schema_version", "tool", "root_size", "widget_count", "widgets"}
-    if not object_fields_are(value, fields, fields):
-        return ["baseline geometry_report has missing or unsupported fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("baseline geometry_report.schema_version must be 1")
-    if value.get("tool") != "FluentQt Named Widget Geometry":
-        errors.append("baseline geometry_report.tool is unsupported")
-    errors.extend(
-        size_report_errors(
-            value.get("root_size"), "baseline geometry_report.root_size"
-        )
-    )
-    widgets = value.get("widgets")
-    if not isinstance(widgets, list):
-        return [*errors, "baseline geometry_report.widgets must be an array"]
-    if not is_json_integer(value.get("widget_count"), minimum=0):
-        errors.append(
-            "baseline geometry_report.widget_count must be a non-negative integer"
-        )
-    elif value.get("widget_count") != len(widgets):
-        errors.append(
-            "baseline geometry_report.widget_count does not match widgets"
-        )
-    widget_fields = {"object_name", "rect"}
-    for index, widget in enumerate(widgets):
-        context = f"baseline geometry_report.widgets[{index}]"
-        if not object_fields_are(widget, widget_fields, widget_fields):
-            errors.append(f"{context} has missing or unsupported fields")
-            continue
-        assert isinstance(widget, dict)
-        if not is_trimmed_nonempty(widget.get("object_name")):
-            errors.append(f"{context}.object_name must be non-empty")
-        errors.extend(rect_report_errors(widget.get("rect"), f"{context}.rect"))
-    return errors
-
-
-def baseline_interaction_report_errors(value: object) -> list[str]:
-    fields = {"schema_version", "requested", "status", "summary", "steps"}
-    if not object_fields_are(value, fields, fields):
-        return ["baseline interaction_report has missing or unsupported fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("baseline interaction_report.schema_version must be 1")
-    if not isinstance(value.get("requested"), bool):
-        errors.append("baseline interaction_report.requested must be a boolean")
-    if value.get("status") not in {"pass", "not-requested"}:
-        errors.append("baseline interaction_report.status is unsupported")
-    summary_fields = {"total", "executed", "passed", "failed"}
-    summary = value.get("summary")
-    if not object_fields_are(summary, summary_fields, summary_fields):
-        errors.append("baseline interaction_report.summary is malformed")
-    elif any(
-        not is_json_integer(summary.get(name), minimum=0)
-        for name in summary_fields
-    ):
-        errors.append("baseline interaction_report.summary values are invalid")
-    steps = value.get("steps")
-    if not isinstance(steps, list):
-        return [*errors, "baseline interaction_report.steps must be an array"]
-    step_fields = {"index", "action", "status"}
-    for index, step in enumerate(steps):
-        context = f"baseline interaction_report.steps[{index}]"
-        if not object_fields_are(step, step_fields, step_fields):
-            errors.append(f"{context} has missing or unsupported fields")
-            continue
-        assert isinstance(step, dict)
-        if not is_json_integer(step.get("index"), minimum=0):
-            errors.append(f"{context}.index must be non-negative")
-        elif step.get("index") != index:
-            errors.append(f"{context}.index does not match its position")
-        if step.get("action") not in ACTION_NAMES:
-            errors.append(f"{context}.action is unsupported")
-        if step.get("status") != "pass":
-            errors.append(f"{context}.status must be pass")
-    if isinstance(summary, dict) and all(
-        is_json_integer(summary.get(name), minimum=0)
-        for name in summary_fields
-    ):
-        if summary.get("total") != len(steps):
-            errors.append(
-                "baseline interaction_report.summary.total does not match steps"
-            )
-        if summary.get("executed") != len(steps):
-            errors.append(
-                "baseline interaction_report.summary.executed does not match steps"
-            )
-        if summary.get("passed") != len(steps) or summary.get("failed") != 0:
-            errors.append(
-                "baseline interaction_report.summary outcomes do not match steps"
-            )
-    if value.get("requested") is True and value.get("status") != "pass":
-        errors.append("requested baseline interaction_report must pass")
-    if value.get("requested") is False and (
-        value.get("status") != "not-requested" or steps
-    ):
-        errors.append("not-requested baseline interaction_report is inconsistent")
-    return errors
-
-
-def baseline_quality_report_errors(value: object) -> list[str]:
-    fields = {"schema_version", "tool", "summary"}
-    if not object_fields_are(value, fields, fields):
-        return ["baseline quality_report has missing or unsupported fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != 1:
-        errors.append("baseline quality_report.schema_version must be 1")
-    if value.get("tool") != "FluentQt Inspector":
-        errors.append("baseline quality_report.tool is unsupported")
-    summary = value.get("summary")
-    summary_fields = {"findings", "by_severity", "by_category"}
-    if not object_fields_are(summary, summary_fields, summary_fields):
-        return [*errors, "baseline quality_report.summary is malformed"]
-    assert isinstance(summary, dict)
-    if not is_json_integer(summary.get("findings"), minimum=0):
-        errors.append("baseline quality_report.summary.findings is invalid")
-    severities = {"info", "warning", "error"}
-    by_severity = summary.get("by_severity")
-    if not object_fields_are(by_severity, severities, severities) or any(
-        not is_json_integer(by_severity.get(name), minimum=0)
-        for name in severities
-    ):
-        errors.append("baseline quality_report.summary.by_severity is malformed")
-    valid_categories = {
-        "text",
-        "accessibility",
-        "input",
-        "focus",
-        "layout",
-        "actions",
-        "scrolling",
-    }
-    by_category = summary.get("by_category")
-    if not isinstance(by_category, dict) or not all(
-        name in valid_categories and is_json_integer(count, minimum=0)
-        for name, count in by_category.items()
-    ):
-        errors.append("baseline quality_report.summary.by_category is malformed")
-    findings = summary.get("findings")
-    if (
-        is_json_integer(findings, minimum=0)
-        and isinstance(by_severity, dict)
-        and all(
-            is_json_integer(by_severity.get(name), minimum=0)
-            for name in severities
-        )
-        and sum(int(by_severity[name]) for name in severities) != findings
-    ):
-        errors.append(
-            "baseline quality_report.summary.by_severity does not match findings"
-        )
-    if (
-        is_json_integer(findings, minimum=0)
-        and isinstance(by_category, dict)
-        and all(
-            name in valid_categories and is_json_integer(count, minimum=0)
-            for name, count in by_category.items()
-        )
-        and sum(int(count) for count in by_category.values()) != findings
-    ):
-        errors.append(
-            "baseline quality_report.summary.by_category does not match findings"
-        )
-    return errors
-
-
-def baseline_report_errors(value: object) -> list[str]:
-    fields = {
-        "schema_version",
-        "tool",
-        "environment_sha256",
-        "geometry_report",
-        "interaction_report",
-        "quality_report",
-    }
-    if not object_fields_are(value, fields, fields):
-        return ["baseline report has missing or unsupported top-level fields"]
-    assert isinstance(value, dict)
-    errors: list[str] = []
-    if value.get("schema_version") != BASELINE_SCHEMA_VERSION:
-        errors.append("baseline report schema_version is unsupported")
-    if value.get("tool") != "FluentQt GUI Baseline Report":
-        errors.append("baseline report tool is unsupported")
-    if not is_sha256(value.get("environment_sha256")):
-        errors.append("baseline report environment_sha256 is malformed")
-    errors.extend(
-        baseline_geometry_report_errors(value.get("geometry_report"))
-    )
-    errors.extend(
-        baseline_interaction_report_errors(value.get("interaction_report"))
-    )
-    errors.extend(baseline_quality_report_errors(value.get("quality_report")))
-    return errors
-
-
-def validated_widget_index(
-    report: Mapping[str, object], *, sanitized: bool = False
-) -> tuple[dict[str, list[dict[str, object]]], list[str]]:
-    geometry = report.get("geometry_report")
-    errors = (
-        baseline_geometry_report_errors(geometry)
-        if sanitized
-        else geometry_report_errors(geometry)
-    )
-    if errors:
-        return {}, errors
-    assert isinstance(geometry, dict)
-    widgets = geometry.get("widgets")
-    index: dict[str, list[dict[str, object]]] = {}
-    assert isinstance(widgets, list)
-    for item in widgets:
-        assert isinstance(item, dict)
-        name = item.get("object_name")
-        if isinstance(name, str) and name:
-            index.setdefault(name, []).append(item)
-    return index, []
-
-
-def normalized_geometry_entry(raw: object, default_tolerance: int) -> dict[str, object]:
-    if isinstance(raw, str):
-        return {"object_name": raw, "tolerance": default_tolerance, "not_clipped": True}
-    if not isinstance(raw, dict):
-        raise VerificationError("geometry.required entries must be strings or objects")
-    result = dict(raw)
-    result.setdefault("tolerance", default_tolerance)
-    result.setdefault("not_clipped", True)
-    return result
-
-
-def rect_deltas(actual: Mapping[str, object], expected: Mapping[str, object]) -> dict[str, int]:
-    return {
-        key: int(actual.get(key, 0)) - int(expected.get(key, 0))
-        for key in ("x", "y", "width", "height")
-    }
-
-
-def geometry_contract_check(report: Mapping[str, object], policy: Mapping[str, object]) -> dict[str, object]:
-    index, report_errors = validated_widget_index(report)
-    if report_errors:
-        return check(
-            "geometry.contract",
-            "incomplete",
-            "Geometry report is malformed and cannot support a contract claim.",
-            {"validation_errors": report_errors},
-        )
-    required = policy.get("required")
-    if not isinstance(required, list) or not required:
-        return check("geometry.contract", "incomplete", "No required geometry probes were declared.")
-    default_tolerance = int(policy.get("tolerance", 0))
-    probes: list[dict[str, object]] = []
-    failures: list[str] = []
-    for raw in required:
-        entry = normalized_geometry_entry(raw, default_tolerance)
-        name = entry.get("object_name")
-        if not isinstance(name, str) or not name:
-            failures.append("geometry probe has no object_name")
-            continue
-        matches = index.get(name, [])
-        probe: dict[str, object] = {"object_name": name, "matches": len(matches)}
-        if len(matches) != 1:
-            failures.append(f"{name} has {len(matches)} matches")
-            probes.append(probe)
-            continue
-        widget = matches[0]
-        probe["actual"] = widget
-        rect = widget.get("rect") if isinstance(widget.get("rect"), dict) else {}
-        visible_rect = (
-            widget.get("visible_rect")
-            if isinstance(widget.get("visible_rect"), dict)
-            else {}
-        )
-        if any(int(rect.get(name, 0)) <= 0 for name in ("width", "height")):
-            failures.append(f"{name} has a zero-size rect")
-        if any(
-            int(visible_rect.get(name, 0)) <= 0
-            for name in ("width", "height")
-        ):
-            failures.append(f"{name} has no visible area")
-        if entry.get("not_clipped", True) and widget.get("clipped") is True:
-            failures.append(f"{name} is clipped")
-        for key, comparison in (("min_width", ">="), ("min_height", ">="), ("max_width", "<="), ("max_height", "<=")):
-            if key not in entry:
-                continue
-            dimension = "width" if "width" in key else "height"
-            actual = int(rect.get(dimension, 0))
-            expected = int(entry[key])
-            bad = actual < expected if comparison == ">=" else actual > expected
-            if bad:
-                failures.append(f"{name}.{dimension} {actual} violates {key}={expected}")
-        expected_rect = entry.get("rect")
-        if isinstance(expected_rect, dict):
-            tolerance = int(entry.get("tolerance", default_tolerance))
-            deltas = rect_deltas(rect, expected_rect)
-            probe["explicit_rect_delta"] = deltas
-            if any(abs(value) > tolerance for value in deltas.values()):
-                failures.append(f"{name} differs from its explicit rect by more than {tolerance}px")
-        probes.append(probe)
-    return check(
-        "geometry.contract",
-        "fail" if failures else "pass",
-        "; ".join(failures) if failures else "All named geometry probes are unique, visible, and within contract.",
-        {"probes": probes},
-    )
-
-
-def geometry_baseline_check(
-    actual_report: Mapping[str, object],
-    baseline_report: Mapping[str, object] | None,
-    policy: Mapping[str, object],
-) -> dict[str, object]:
-    if baseline_report is None:
-        return check(
-            "geometry.baseline",
-            "human-required",
-            "Approved baseline geometry is not available.",
-        )
-    actual_index, actual_errors = validated_widget_index(actual_report)
-    baseline_index, baseline_errors = validated_widget_index(
-        baseline_report, sanitized=True
-    )
-    if actual_errors or baseline_errors:
-        return check(
-            "geometry.baseline",
-            "incomplete",
-            "Actual or approved geometry report is malformed.",
-            {
-                "actual_validation_errors": actual_errors,
-                "baseline_validation_errors": baseline_errors,
-            },
-        )
-    required = policy.get("required") if isinstance(policy.get("required"), list) else []
-    default_tolerance = int(policy.get("tolerance", 0))
-    comparisons: list[dict[str, object]] = []
-    failures: list[str] = []
-    required_names = [
-        str(normalized_geometry_entry(raw, default_tolerance)["object_name"])
-        for raw in required
-    ]
-    baseline_geometry = baseline_report.get("geometry_report")
-    assert isinstance(baseline_geometry, dict)
-    baseline_widgets = baseline_geometry.get("widgets")
-    assert isinstance(baseline_widgets, list)
-    baseline_names = [
-        str(widget.get("object_name", "")) for widget in baseline_widgets
-    ]
-    if baseline_names != required_names:
-        failures.append(
-            "approved baseline geometry must contain exactly the recipe-required "
-            "probes in declared order"
-        )
-    for raw in required:
-        entry = normalized_geometry_entry(raw, default_tolerance)
-        name = str(entry.get("object_name", ""))
-        actual = actual_index.get(name, [])
-        baseline = baseline_index.get(name, [])
-        if len(actual) != 1 or len(baseline) != 1:
-            failures.append(
-                f"{name} must be unique in actual and baseline ({len(actual)}/{len(baseline)})"
-            )
-            continue
-        actual_rect = actual[0].get("rect") if isinstance(actual[0].get("rect"), dict) else {}
-        baseline_rect = baseline[0].get("rect") if isinstance(baseline[0].get("rect"), dict) else {}
-        deltas = rect_deltas(actual_rect, baseline_rect)
-        tolerance = int(entry.get("tolerance", default_tolerance))
-        comparisons.append(
-            {
-                "object_name": name,
-                "actual": actual_rect,
-                "baseline": baseline_rect,
-                "delta": deltas,
-                "tolerance": tolerance,
-            }
-        )
-        if any(abs(value) > tolerance for value in deltas.values()):
-            failures.append(f"{name} moved or resized by more than {tolerance}px")
-    return check(
-        "geometry.baseline",
-        "fail" if failures else "pass",
-        "; ".join(failures) if failures else "Named geometry matches the approved baseline.",
-        {"comparisons": comparisons},
     )
 
 
@@ -3453,248 +1101,6 @@ def native_desktop_check(report: Mapping[str, object], required: bool) -> dict[s
     )
 
 
-def comparator_policy_arguments(policy: Mapping[str, object]) -> list[str]:
-    mapping = (
-        ("channel_threshold", "--channel-threshold"),
-        ("max_different_pixels", "--max-different-pixels"),
-        ("max_different_ratio", "--max-different-ratio"),
-        ("search_radius", "--search-radius"),
-        ("max_translation", "--max-translation"),
-        ("edge_threshold", "--edge-threshold"),
-    )
-    arguments: list[str] = []
-    for key, flag in mapping:
-        if key in policy:
-            arguments.extend([flag, str(policy[key])])
-    return arguments
-
-
-def expected_comparator_policy(policy: Mapping[str, object]) -> dict[str, object]:
-    max_pixels = policy.get("max_different_pixels")
-    max_ratio = policy.get("max_different_ratio")
-    if max_pixels is None and max_ratio is None:
-        max_pixels = 0
-    return {
-        "channel_threshold": policy.get("channel_threshold", 0),
-        "max_different_pixels": max_pixels,
-        "max_different_ratio": max_ratio,
-        "translation_search_radius": policy.get("search_radius", 4),
-        "max_translation": policy.get("max_translation", 0),
-        "edge_threshold": policy.get("edge_threshold", 12),
-    }
-
-
-def expected_region(region: list[int] | None) -> dict[str, int] | None:
-    if region is None:
-        return None
-    return dict(zip(("x", "y", "width", "height"), region))
-
-
-def validate_comparator_report(
-    report: Mapping[str, object],
-    returncode: int,
-    baseline_path: Path,
-    actual_path: Path,
-    region: list[int] | None,
-    policy: Mapping[str, object],
-) -> list[str]:
-    errors: list[str] = []
-    baseline_dimensions = png_dimensions(baseline_path)
-    actual_dimensions = png_dimensions(actual_path)
-    if baseline_dimensions is None:
-        errors.append("comparator baseline is not a valid PNG")
-    if actual_dimensions is None:
-        errors.append("comparator actual is not a valid PNG")
-    expected_dimensions: tuple[int, int] | None = None
-    if baseline_dimensions is not None and actual_dimensions is not None:
-        if region is None:
-            if baseline_dimensions != actual_dimensions:
-                errors.append("comparator input PNG dimensions do not match")
-            else:
-                expected_dimensions = baseline_dimensions
-        else:
-            x, y, width, height = region
-            expected_dimensions = (width, height)
-            for label, dimensions in (
-                ("baseline", baseline_dimensions),
-                ("actual", actual_dimensions),
-            ):
-                if (
-                    x < 0
-                    or y < 0
-                    or x + width > dimensions[0]
-                    or y + height > dimensions[1]
-                ):
-                    errors.append(
-                        f"comparator region is outside the {label} PNG bounds"
-                    )
-    if report.get("schema_version") != 1:
-        errors.append("unsupported comparator report schema")
-    if report.get("tool") != "FluentQt Visual Compare":
-        errors.append("unexpected comparator tool identity")
-    expected_status = "pass" if returncode == 0 else "fail" if returncode == 1 else None
-    if expected_status is None or report.get("status") != expected_status:
-        errors.append("comparator status does not match its process return code")
-    inputs = report.get("inputs") if isinstance(report.get("inputs"), dict) else {}
-    if Path(str(inputs.get("baseline", ""))).resolve() != baseline_path.resolve():
-        errors.append("comparator baseline input does not match the requested file")
-    if Path(str(inputs.get("actual", ""))).resolve() != actual_path.resolve():
-        errors.append("comparator actual input does not match the requested file")
-    if inputs.get("baseline_sha256") != sha256_file(baseline_path):
-        errors.append("comparator baseline digest is missing or stale")
-    if inputs.get("actual_sha256") != sha256_file(actual_path):
-        errors.append("comparator actual digest is missing or stale")
-    if inputs.get("region") != expected_region(region):
-        errors.append("comparator region does not match the requested crop")
-    expected_size = (
-        {"width": expected_dimensions[0], "height": expected_dimensions[1]}
-        if expected_dimensions is not None
-        else None
-    )
-    if report.get("baseline_size") != expected_size:
-        errors.append("comparator baseline size does not match the input PNG")
-    if report.get("actual_size") != expected_size:
-        errors.append("comparator actual size does not match the input PNG")
-    if report.get("policy") != expected_comparator_policy(policy):
-        errors.append("comparator report policy does not match the requested policy")
-    checks = report.get("checks") if isinstance(report.get("checks"), dict) else {}
-    required_checks = {
-        "size_matches",
-        "pixel_limits_pass",
-        "translation_limit_pass",
-    }
-    if set(checks) != required_checks or not all(
-        isinstance(checks.get(name), bool) for name in required_checks
-    ):
-        errors.append("comparator checks are missing or malformed")
-    elif expected_status == "pass" and not all(checks.values()):
-        errors.append("passing comparator report contains a failed check")
-    elif expected_status == "fail" and all(checks.values()):
-        errors.append("failing comparator report contains no failed check")
-    metrics = report.get("metrics") if isinstance(report.get("metrics"), dict) else {}
-    total = metrics.get("total_pixels")
-    different = metrics.get("different_pixels")
-    ratio = metrics.get("different_ratio")
-    numeric = lambda value: isinstance(value, (int, float)) and not isinstance(value, bool)
-    if not is_json_integer(total, minimum=1):
-        errors.append("comparator total_pixels must be a positive integer")
-    elif expected_dimensions is not None and total != (
-        expected_dimensions[0] * expected_dimensions[1]
-    ):
-        errors.append("comparator total_pixels does not match the compared image size")
-    if not is_json_integer(different, minimum=0) or (
-        is_json_integer(total, minimum=1) and different > total
-    ):
-        errors.append("comparator different_pixels is outside the image bounds")
-    if not numeric(ratio) or not 0 <= ratio <= 1:
-        errors.append("comparator different_ratio is outside 0..1")
-    elif is_json_integer(total, minimum=1) and is_json_integer(different, minimum=0):
-        expected_ratio = different / total
-        if abs(ratio - expected_ratio) > 1e-9:
-            errors.append("comparator different_ratio does not match pixel counts")
-    expected_policy = expected_comparator_policy(policy)
-    if is_json_integer(different, minimum=0) and numeric(ratio):
-        max_pixels = expected_policy["max_different_pixels"]
-        max_ratio = expected_policy["max_different_ratio"]
-        expected_pixel_limit = (
-            (max_pixels is None or different <= max_pixels)
-            and (max_ratio is None or ratio <= max_ratio)
-        )
-        if checks.get("pixel_limits_pass") != expected_pixel_limit:
-            errors.append(
-                "comparator pixel limit check does not match its metrics"
-            )
-    translation = (
-        metrics.get("estimated_translation")
-        if isinstance(metrics.get("estimated_translation"), dict)
-        else {}
-    )
-    translation_fields = {
-        "dx",
-        "dy",
-        "confident",
-        "baseline_edge_pixels",
-        "actual_edge_pixels",
-        "zero_offset_score",
-        "best_score",
-        "improvement",
-    }
-    if set(translation) != translation_fields:
-        errors.append("comparator estimated_translation is missing or malformed")
-    else:
-        for name in ("dx", "dy"):
-            if not is_json_integer(translation.get(name)):
-                errors.append(
-                    f"comparator estimated_translation.{name} must be an integer"
-                )
-        if not isinstance(translation.get("confident"), bool):
-            errors.append(
-                "comparator estimated_translation.confident must be a boolean"
-            )
-        for name in ("baseline_edge_pixels", "actual_edge_pixels"):
-            if not is_json_integer(translation.get(name), minimum=0):
-                errors.append(
-                    f"comparator estimated_translation.{name} must be a non-negative integer"
-                )
-        for name in ("zero_offset_score", "best_score", "improvement"):
-            if not numeric(translation.get(name)):
-                errors.append(
-                    f"comparator estimated_translation.{name} must be numeric"
-                )
-    max_translation = expected_policy["max_translation"]
-    if max_translation is not None and is_json_integer(different, minimum=0):
-        if different == 0 or translation.get("confident") is not True:
-            expected_translation_limit = True
-        else:
-            dx = translation.get("dx")
-            dy = translation.get("dy")
-            expected_translation_limit = (
-                translation.get("confident") is True
-                and is_json_integer(dx)
-                and is_json_integer(dy)
-                and abs(dx) <= max_translation
-                and abs(dy) <= max_translation
-            )
-        if checks.get("translation_limit_pass") != expected_translation_limit:
-            errors.append(
-                "comparator translation check does not match its metrics"
-            )
-    return errors
-
-
-def pixel_comparisons(
-    policy: Mapping[str, object], device_pixel_ratio: float
-) -> list[tuple[str, Mapping[str, object], list[int] | None]]:
-    comparisons: list[tuple[str, Mapping[str, object], list[int] | None]] = [
-        ("full", policy, None)
-    ]
-    regions = policy.get("regions")
-    if not isinstance(regions, list):
-        return comparisons
-    for index, raw in enumerate(regions):
-        if not isinstance(raw, dict):
-            continue
-        region_id = str(raw.get("id", f"region-{index + 1}"))
-        if not SAFE_ID.fullmatch(region_id):
-            region_id = f"region-{index + 1}"
-        rect = raw.get("rect")
-        if not isinstance(rect, list) or len(rect) != 4:
-            continue
-        physical_rect = [int(value) for value in rect]
-        if raw.get("coordinate_space", "logical") == "logical":
-            scaled_rect = [
-                qt_scale_positive(value, device_pixel_ratio)
-                for value in physical_rect
-            ]
-            if any(value is None for value in scaled_rect):
-                continue
-            physical_rect = [int(value) for value in scaled_rect]
-        comparisons.append(
-            (region_id, merged_dict(policy, raw.get("policy")), physical_rect)
-        )
-    return comparisons
-
-
 def reset_expected_output(path: Path, owner_directory: Path) -> None:
     """Remove only a declared producer output so stale evidence cannot be reused."""
 
@@ -4028,60 +1434,6 @@ def run_scenario(
         "pixel_executions": pixel_executions,
         "artifacts": artifacts,
     }
-
-
-def html_uri(path: object) -> str:
-    if not isinstance(path, str) or not path:
-        return ""
-    candidate = Path(path)
-    return candidate.resolve().as_uri() if candidate.exists() else ""
-
-
-def review_html(evidence: Mapping[str, object]) -> str:
-    cards: list[str] = []
-    scenarios = evidence.get("scenarios") if isinstance(evidence.get("scenarios"), list) else []
-    for raw in scenarios:
-        if not isinstance(raw, dict):
-            continue
-        artifacts = raw.get("artifacts") if isinstance(raw.get("artifacts"), dict) else {}
-        images: list[str] = []
-        for label, key in (("Actual", "actual"), ("Approved baseline", "baseline")):
-            uri = html_uri(artifacts.get(key))
-            if uri:
-                images.append(
-                    f'<figure><figcaption>{escape(label)}</figcaption><img src="{escape(uri)}" alt="{escape(label)}"></figure>'
-                )
-        diffs = artifacts.get("diffs") if isinstance(artifacts.get("diffs"), list) else []
-        for index, diff in enumerate(diffs):
-            uri = html_uri(diff)
-            if uri:
-                images.append(
-                    f'<figure><figcaption>Diff {index + 1}</figcaption><img src="{escape(uri)}" alt="Diff"></figure>'
-                )
-        failed_checks = [
-            item
-            for item in raw.get("checks", [])
-            if isinstance(item, dict) and item.get("status") not in {"pass", "not-applicable"}
-        ]
-        check_items = "".join(
-            f'<li><code>{escape(str(item.get("id")))}</code> — {escape(str(item.get("status")))}: {escape(str(item.get("message")))}</li>'
-            for item in failed_checks
-        ) or "<li>All deterministic checks passed.</li>"
-        prompts = "".join(f"<li>{escape(str(prompt))}</li>" for prompt in raw.get("review", []))
-        cards.append(
-            f'''<section>
-<h2>{escape(str(raw.get("id")))} <span class="status {escape(str(raw.get("status")))}">{escape(str(raw.get("status")))}</span></h2>
-<div class="images">{"".join(images)}</div>
-<h3>Deterministic gates</h3><ul>{check_items}</ul>
-<h3>Independent review prompts</h3><ul>{prompts}</ul>
-</section>'''
-        )
-    return f'''<!doctype html>
-<html lang="en"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1">
-<title>FluentQt GUI verification</title>
-<style>
-body{{font:14px system-ui;margin:24px;background:#111318;color:#f4f6fa}}h1{{font-size:22px}}section{{margin:20px 0;padding:16px;border:1px solid #3a3f49;border-radius:10px;background:#191c22}}h2{{margin-top:0}}h3{{font-size:14px}}.images{{display:grid;grid-template-columns:repeat(auto-fit,minmax(300px,1fr));gap:12px}}figure{{margin:0;background:#fff;border-radius:6px;overflow:hidden}}figcaption{{padding:8px;background:#2a2f38;color:#fff}}img{{display:block;width:100%;height:auto;image-rendering:auto}}.status{{font-size:12px;padding:3px 7px;border-radius:999px;background:#404754}}.pass{{background:#176b42}}.fail{{background:#a73333}}.human-required,.review-required,.incomplete{{background:#8a641d}}code{{color:#9fd2ff}}li{{margin:5px 0}}
-</style></head><body><h1>FluentQt GUI verification</h1><p>Deterministic status: <strong>{escape(str(evidence.get("deterministic_status")))}</strong>. Final visual acceptance requires a separate reviewer whose identity differs from the author.</p>{"".join(cards)}</body></html>'''
 
 
 def write_review_request(evidence_path: Path, evidence: Mapping[str, object], output: Path) -> dict[str, object]:
