+++ tools/dev/gui_verification/__init__.py
@@ -1,0 +1,1 @@
+"""Internal modules for the fluent_qt_gui_verify command."""
